### 2021-09-16 #82 @whedon
[PRE REVIEW]: MutableArithmetics: An API for mutable operations
**Submitting author:** @blegat (<a href="http://orcid.org/0000-0002-5921-368X">Benoît Legat</a>)
**Repository:** <a href="https://github.com/jump-dev/MutableArithmetics.jl/" target ="_blank">https://github.com/jump-dev/MutableArithmetics.jl/</a>
**Version:**
