### 2021-09-19 d04483f
Fix paper compilation
