+++ paper/ref.bib
@@ -1,6 +1,7 @@
 @article{bezanson2017julia,
   title={Julia: A fresh approach to numerical computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  doi={10.1137/141000671},
   journal={SIAM review},
   volume={59},
   number={1},
