### 2021-10-11 57b2f6c
Force specialization on mutable operate
