### 2021-10-12 dd62602
Force specialization for all functions
