+++ src/interface.jl
@@ -16,10 +16,10 @@ function promote_operation_fallback(::typeof(/), ::Type{S}, ::Type{T}) where {S,
 end
 # Julia v1.0.x has trouble with inference with the `Vararg` method, see
 # https://travis-ci.org/jump-dev/JuMP.jl/jobs/617606373
-function promote_operation_fallback(op::Function, ::Type{S}, ::Type{T}) where {S,T}
+function promote_operation_fallback(op::F, ::Type{S}, ::Type{T}) where {F<:Function,S,T}
     return typeof(op(zero(S), zero(T)))
 end
-function promote_operation_fallback(op::Function, args::Vararg{Type,N}) where {N}
+function promote_operation_fallback(op::F, args::Vararg{Type,N}) where {F<:Function,N}
     return typeof(op(zero.(args)...))
 end
 promote_operation_fallback(::typeof(*), ::Type{T}) where {T} = T
@@ -54,7 +54,7 @@ end
 Returns the type returned to the call `operate(op, args...)` where the types of
 the arguments `args` are `ArgsTypes`.
 """
-function promote_operation(op::Function, args::Vararg{Type,N}) where {N}
+function promote_operation(op::F, args::Vararg{Type,N}) where {F<:Function,N}
     return promote_operation_fallback(op, args...)
 end
 
@@ -248,7 +248,7 @@ terms of `p` and `q` hence it will never terminate. On the other hand
 `mutable_operate!(+, p, q)` uses a different algorithm that efficiently inserts
 the terms of `q` in the sorted list of terms of `p` with minimal displacement.
 """
-function mutable_operate_to!(output, op::Function, args::Vararg{Any,N}) where {N}
+function mutable_operate_to!(output, op::F, args::Vararg{Any,N}) where {F<:Function,N}
     mutable_operate_to_fallback(mutability(output, op, args...), output, op, args...)
 end
 
@@ -281,7 +281,7 @@ function mutable_operate!(op::F, args::Vararg{Any,N}) where {F<:Function,N}
     mutable_operate_fallback(mutability(args[1], op, args...), op, args...)
 end
 
-buffer_for(::Function, args::Vararg{Type,N}) where {N} = nothing
+buffer_for(::F, args::Vararg{Type,N}) where {F<:Function,N} = nothing
 
 function mutable_buffered_operate_to_fallback(::NotMutable, buffer, output, op::Function, args...)
     throw(
@@ -303,9 +303,9 @@ end
 function mutable_buffered_operate_to_fallback(
     buffer,
     output,
-    op::Function,
+    op::F,
     args::Vararg{Any,N},
-) where {N}
+) where {F<:Function,N}
     return mutable_buffered_operate_to_fallback(
         mutability(output, op, args...),
         buffer,
@@ -318,9 +318,9 @@ end
 function mutable_buffered_operate_to_fallback(
     ::Nothing,
     output,
-    op::Function,
+    op::F,
     args::Vararg{Any,N},
-) where {N}
+) where {F<:Function,N}
     return mutable_operate_to!(output, op, args...)
 end
 
@@ -334,9 +334,9 @@ possibly modifying `buffer`. Can only be called if
 function mutable_buffered_operate_to!(
     buffer,
     output,
-    op::Function,
+    op::F,
     args::Vararg{Any,N},
-) where {N}
+) where {F<:Function,N}
     return mutable_buffered_operate_to_fallback(buffer, output, op, args...)
 end
 
@@ -358,9 +358,9 @@ end
 
 function mutable_buffered_operate_fallback(
     buffer,
-    op::Function,
+    op::F,
     args::Vararg{Any,N},
-) where {N}
+) where {F<:Function,N}
[…]
