### 2021-10-12 4462827
Fix call of promote_operation in @rewrite (#119)
