### 2021-10-16 #93 @whedon
[REVIEW]: MutableArithmetics: An API for mutable operations
**Submitting author:** @blegat (<a href="http://orcid.org/0000-0002-5921-368X">Benoît Legat</a>)
**Repository:** https://github.com/jump-dev/MutableArithmetics.jl/
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.3
