### 2021-10-16 4fe8bdb
Force specialization for all functions (#120)
