+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#","page":"Home","title":"Home","text":"CurrentModule = MutableArithmetics\nDocTestSetup = quote\n    using MutableArithmetics\nend","category":"page"},{"location":"#MutableArithmetics.jl-1","page":"Home","title":"MutableArithmetics.jl","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"","category":"page"},{"location":"#","page":"Home","title":"Home","text":"Modules = [MutableArithmetics]","category":"page"},{"location":"#MutableArithmetics.@rewrite-Tuple{Any}","page":"Home","title":"MutableArithmetics.@rewrite","text":"@rewrite(expr)\n\nReturn the value of expr exploiting the mutability of the temporary expressions created for the computation of the result.\n\nExamples\n\nThe expression\n\nMA.@rewrite(x + y * z + u * v * w)\n\nis rewritten into\n\nMA.add_mul!(MA.add_mul!(MA.copy_if_mutable(x),\n                        y, z),\n            u, v, w)\n\n\n\n\n\n","category":"macro"},{"location":"#MutableArithmetics.IsMutable","page":"Home","title":"MutableArithmetics.IsMutable","text":"struct IsMutable <: MutableTrait end\n\nWhen this is returned by mutability, it means that object of the given type can always be mutated to equal the result of the operation.\n\n\n\n\n\n","category":"type"},{"location":"#MutableArithmetics.MutableTrait","page":"Home","title":"MutableArithmetics.MutableTrait","text":"abstract type MutableTrait end\n\nAbstract type for IsMutable and NotMutable that are returned by mutability.\n\n\n\n\n\n","category":"type"},{"location":"#MutableArithmetics.NotMutable","page":"Home","title":"MutableArithmetics.NotMutable","text":"struct NotMutable <: MutableTrait end\n\nWhen this is returned by mutability, it means that object of the given type cannot be mutated to equal the result of the operation.\n\n\n\n\n\n","category":"type"},{"location":"#MutableArithmetics._rewrite","page":"Home","title":"MutableArithmetics._rewrite","text":"_rewrite(\n    vectorized::Bool,\n    minus::Bool,\n    inner_factor,\n    current_sum::Union{Symbol, Nothing},\n    left_factors::Vector,\n    right_factors::Vector,\n    new_var::Symbol = gensym(),\n)\n\nReturn new_var, code such that code is equivalent to\n\nnew_var = prod(left_factors) * inner_factor * prod(reverse(right_factors))\n\nIf current_sum is nothing, and is\n\nnew_var = current_sum op prod(left_factors) * inner_factor * prod(reverse(right_factors))\n\notherwise where op is + if !vectorized & !minus, .+ if vectorized & !minus, - if !vectorized & minus and .- if vectorized & minus.\n\n\n\n\n\n","category":"function"},{"location":"#MutableArithmetics.add!-Union{Tuple{Vararg{Any,N}}, Tuple{N}} where N","page":"Home","title":"MutableArithmetics.add!","text":"add!(a, b, ...)\n\nReturn the sum of a, b, ..., possibly modifying a.\n\n\n\n\n\n","category":"method"},{"location":"#MutableArithmetics.add_dot-Union{Tuple{N}, Tuple{Any,Any,Any,Vararg{Any,N}}} where N","page":"Home","title":"MutableArithmetics.add_dot","text":"add_dot(a, args...)\n\nReturn a + dot(args...).\n\n\n\n\n\n","category":"method"},{"location":"#MutableArithmetics.add_mul","page":"Home","title":"MutableArithmetics.add_mul","text":"add_mu
[…]
