+++ src/dispatch.jl
@@ -213,86 +213,86 @@ end
 # A few are overwritten below but many more need to be redirected to `mul` in
 # `linalg.jl`.
 
-Base.:*(A::SparseMat{<:AbstractMutable}, x::StridedVector) = mul(A, x)
-Base.:*(A::SparseMat, x::StridedVector{<:AbstractMutable}) = mul(A, x)
-Base.:*(A::SparseMat{<:AbstractMutable}, x::StridedVector{<:AbstractMutable}) = mul(A, x)
+Base.:*(A::_SparseMat{<:AbstractMutable}, x::StridedVector) = mul(A, x)
+Base.:*(A::_SparseMat, x::StridedVector{<:AbstractMutable}) = mul(A, x)
+Base.:*(A::_SparseMat{<:AbstractMutable}, x::StridedVector{<:AbstractMutable}) = mul(A, x)
 
 # These six methods are needed on Julia v1.2 and earlier
-Base.:*(A::LinearAlgebra.Adjoint{<:AbstractMutable,<:SparseMat}, x::StridedVector) =
+Base.:*(A::LinearAlgebra.Adjoint{<:AbstractMutable,<:_SparseMat}, x::StridedVector) =
     mul(A, x)
-Base.:*(A::LinearAlgebra.Adjoint{<:Any,<:SparseMat}, x::StridedVector{<:AbstractMutable}) =
-    mul(A, x)
-Base.:*(
-    A::LinearAlgebra.Adjoint{<:AbstractMutable,<:SparseMat},
-    x::StridedVector{<:AbstractMutable},
-) = mul(A, x)
-Base.:*(A::LinearAlgebra.Transpose{<:AbstractMutable,<:SparseMat}, x::StridedVector) =
+Base.:*(A::LinearAlgebra.Adjoint{<:Any,<:_SparseMat}, x::StridedVector{<:AbstractMutable}) =
     mul(A, x)
 Base.:*(
-    A::LinearAlgebra.Transpose{<:Any,<:SparseMat},
+    A::LinearAlgebra.Adjoint{<:AbstractMutable,<:_SparseMat},
+    x::StridedVector{<:AbstractMutable},
+) = mul(A, x)
+Base.:*(A::LinearAlgebra.Transpose{<:AbstractMutable,<:_SparseMat}, x::StridedVector) =
+    mul(A, x)
+Base.:*(
+    A::LinearAlgebra.Transpose{<:Any,<:_SparseMat},
     x::StridedVector{<:AbstractMutable},
 ) = mul(A, x)
 Base.:*(
-    A::LinearAlgebra.Transpose{<:AbstractMutable,<:SparseMat},
+    A::LinearAlgebra.Transpose{<:AbstractMutable,<:_SparseMat},
     x::StridedVector{<:AbstractMutable},
 ) = mul(A, x)
 
-Base.:*(A::SparseMat{<:AbstractMutable}, B::SparseMat{<:AbstractMutable}) = mul(A, B)
-Base.:*(A::SparseMat{<:Any}, B::SparseMat{<:AbstractMutable}) = mul(A, B)
-Base.:*(A::SparseMat{<:AbstractMutable}, B::SparseMat{<:Any}) = mul(A, B)
+Base.:*(A::_SparseMat{<:AbstractMutable}, B::_SparseMat{<:AbstractMutable}) = mul(A, B)
+Base.:*(A::_SparseMat{<:Any}, B::_SparseMat{<:AbstractMutable}) = mul(A, B)
+Base.:*(A::_SparseMat{<:AbstractMutable}, B::_SparseMat{<:Any}) = mul(A, B)
 
 Base.:*(
-    A::SparseMat{<:AbstractMutable},
-    B::LinearAlgebra.Adjoint{<:AbstractMutable,<:SparseMat},
[…]
