### 2021-10-17 fe6b8f5
NotMutable -> IsNotMutable (#118)
* NotMutable -> IsNotMutable

* Update tests
