+++ src/MutableArithmetics.jl
@@ -191,7 +191,6 @@ for sym in names(@__MODULE__, all = true)
     )
         continue
     end
-    @show sym
     @eval export $sym
 end
 
