### 2021-10-17 4f4cd07
Fix buffer_for for Rational
