### 2021-10-17 17c93f9
Implement MutableArithmetics for Rational (#104)
* Implement MutableArithmetics for Rational

* Fixes

* Fix

* Fix allocation issue

* Skip test for Julia v1.0

* Fix buffer_for for Rational
