+++ src/rational.jl
@@ -1,0 +1,159 @@
+# `Rational` is a `struct`, not a `mutable struct` so if `T` is not
+# mutable, we cannot mutate the rational.
+mutability(::Type{Rational{T}}) where {T} = mutability(T)
+mutable_copy(x::Rational) = Rational(mutable_copy(x.num), mutable_copy(x.den))
+
+# zero
+promote_operation(::typeof(zero), ::Type{Rational{T}}) where {T} = Rational{T}
+function mutable_operate!(::typeof(zero), x::Rational)
+    mutable_operate!(zero, x.num)
+    mutable_operate!(one, x.den)
+    return x
+end
+
+# one
+promote_operation(::typeof(one), ::Type{Rational{T}}) where {T} = Rational{T}
+function mutable_operate!(::typeof(one), x::Rational)
+    mutable_operate!(one, x.num)
+    mutable_operate!(one, x.den)
+    return x
+end
+
+# +
+function promote_operation(::typeof(+), ::Type{Rational{S}}, ::Type{Rational{T}}) where {S,T}
+    return Rational{promote_sum_mul(S, T)}
+end
+function mutable_operate_to!(output::Rational, ::typeof(+), x::Rational, y::Rational)
+    xd, yd = Base.divgcd(promote(x.den, y.den)...)
+    # TODO Use `checked_mul` and `checked_add` like in Base
+    mutable_operate_to!(output.num, *, x.num, yd)
+    mutable_operate!(add_mul, output.num, y.num, xd)
+    mutable_operate_to!(output.den, *, x.den, yd)
+    return output
+end
+
+# -
+function promote_operation(::typeof(-), ::Type{Rational{S}}, ::Type{Rational{T}}) where {S,T}
+    return Rational{promote_sum_mul(S, T)}
+end
+function mutable_operate_to!(output::Rational, ::typeof(-), x::Rational, y::Rational)
+    xd, yd = Base.divgcd(promote(x.den, y.den)...)
+    # TODO Use `checked_mul` and `checked_sub` like in Base
+    mutable_operate_to!(output.num, *, x.num, yd)
+    mutable_operate!(sub_mul, output.num, y.num, xd)
+    mutable_operate_to!(output.den, *, x.den, yd)
+    return output
+end
+
+# *
+function promote_operation(::typeof(*), ::Type{Rational{S}}, ::Type{Rational{T}}) where {S,T}
+    return Rational{promote_operation(*, S, T)}
+end
+function mutable_operate_to!(output::Rational, ::typeof(*), x::Rational, y::Rational)
+    xn, yd = Base.divgcd(promote(x.num, y.den)...)
+    xd, yn = Base.divgcd(promote(x.den, y.num)...)
+    mutable_operate_to!(output.num, *, xn, yn)
+    mutable_operate_to!(output.den, *, xd, yd)
+    return output
+end
+
+# gcd
+function promote_operation(::Union{typeof(gcd),typeof(lcm)}, ::Type{Rational{S}}, ::Type{Rational{T}}) where {S,T}
+    return Rational{promote_operation(gcd, S, T)}
+end
+function mutable_operate_to!(output::Rational, ::typeof(gcd), a::Rational, b::Rational)
+    mutable_operate_to!(output.num, gcd, a.num, b.num)
+    mutable_operate_to!(output.den, lcm, a.den, b.den)
+    return output
+end
+function mutable_operate_to!(output::Rational, ::typeof(lcm), a::Rational, b::Rational)
+    mutable_operate_to!(output.num, lcm, a.num, b.num)
+    mutable_operate_to!(output.den, gcd, a.den, b.den)
+    return output
+end
+
+function mutable_operate_to!(
+    output::Rational,
[…]
