### 2021-10-17 a146c3a
Renaming following BangBang.jl convention
