+++ src/interface.jl
@@ -1,7 +1,7 @@
 # Example of mutable types that can implement this API: BigInt, Array, JuMP.AffExpr, MultivariatePolynomials.AbstractPolynomial
-# `mutable_operate!(add_mul, ...)` is similar to `JuMP.add_to_expression(...)`
-# `operate!(add_mul, ...)` is similar to `JuMP.destructive_add(...)`
-# `operate!` is similar to `MOI.Utilities.operate!`
+# `operate!(add_mul, ...)` is similar to `JuMP.add_to_expression(...)`
+# `operate!!(add_mul, ...)` is similar to `JuMP.destructive_add(...)`
+# `operate!!` is similar to `MOI.Utilities.operate!`
 
 # `promote_operation_fallback` gives fallbacks for any type with no risk of
 # ambiguity with specific methods defined for a given type, even if these are
@@ -213,100 +213,100 @@ implement a method for [`mutable_copy`](@ref) instead.
 """
 copy_if_mutable(x) = copy_if_mutable_fallback(mutability(typeof(x)), x)
 
-function mutable_operate_to_fallback(::IsNotMutable, output, op::Function, args...)
+function operate_to_fallback!(::IsNotMutable, output, op::Function, args...)
     throw(
         ArgumentError(
-            "Cannot call `mutable_operate_to!(::$(typeof(output)), $op, ::$(join(typeof.(args), ", ::")))` as objects of type `$(typeof(output))` cannot be modifed to equal the result of the operation. Use `operate_to!` instead which returns the value of the result (possibly modifying the first argument) to write generic code that also works when the type cannot be modified.",
+            "Cannot call `operate_to!(::$(typeof(output)), $op, ::$(join(typeof.(args), ", ::")))` as objects of type `$(typeof(output))` cannot be modifed to equal the result of the operation. Use `operate_to!!` instead which returns the value of the result (possibly modifying the first argument) to write generic code that also works when the type cannot be modified.",
         ),
     )
 end
 
-function mutable_operate_to_fallback(::IsMutable, output, op::AddSubMul, x, y)
-    return mutable_operate_to!(output, add_sub_op(op), x, y)
+function operate_to_fallback!(::IsMutable, output, op::AddSubMul, x, y)
+    return operate_to!(output, add_sub_op(op), x, y)
 end
-function mutable_operate_to_fallback(::IsMutable, output, op::Function, args...)
+function operate_to_fallback!(::IsMutable, output, op::Function, args...)
     error(
-        "`mutable_operate_to!(::$(typeof(output)), $op, ::",
+        "`operate_to!(::$(typeof(output)), $op, ::",
         join(typeof.(args), ", ::"),
         ")` is not implemented yet.",
     )
 end
 
 """
-    mutable_operate_to!(output, op::Function, args...)
+    operate_to!(output, op::Function, args...)
 
 Modify the value of `output` to be equal to the value of `op(args...)`. Can
-only be called if `mutability(output, op, args...)` returns `true`.
+only be called if `mutability(output, op, args...)` returns [`IsMutable`](@ref).
 
 If `output === args[i]` for some `i`,
-* The user should expect to get an error. `operate!` or `mutable_operate!` should be used instead.
+* The user should expect to get an error. `operate!!` or `operate!` should be used instead.
 * Any method not supporting this case should throw an error.
 
-For instance, in DynamicPolynomials, `mutable_operate_to!(p, +, p, q)` throws an
+For instance, in DynamicPolynomials, `operate_to!(p, +, p, q)` throws an
[…]
