+++ src/MutableArithmetics.jl
@@ -152,31 +152,6 @@ include("dispatch.jl")
 # Test that can be used to test an implementation of the interface
 include("Test/Test.jl")
 
-# JuMP exports everything except internal symbols, which are defined as those
-# whose name starts with an underscore. Macros whose names start with
-# underscores are internal as well. If you don't want all of these symbols
-# in your environment, then use `import JuMP` instead of `using JuMP`.
-
-# Do not add JuMP-defined symbols to this exclude list. Instead, rename them
-# with an underscore.
-const _EXCLUDE_SYMBOLS = [Symbol(@__MODULE__), :eval, :include]
-
-for sym in names(@__MODULE__, all = true)
-    sym_string = string(sym)
-    if sym in _EXCLUDE_SYMBOLS ||
-       startswith(sym_string, "_") ||
-       startswith(sym_string, "@_")
-        continue
-    end
-    if !(
-        Base.isidentifier(sym) ||
-        (startswith(sym_string, "@") && Base.isidentifier(sym_string[2:end]))
-    )
-        continue
-    end
-    @eval export $sym
-end
-
 # This export code below was taken from JuMP:
 const _EXCLUDE_SYMBOLS = [Symbol(@__MODULE__), :eval, :include]
 for sym in names(@__MODULE__, all = true)
