### 2021-10-19 2379436
Renaming following BangBang.jl convention (#121)
