+++ src/linear_algebra.jl
@@ -35,26 +35,26 @@ function promote_operation(
 end
 
 # Only `Scaling`
-function mutable_operate!(
+function operate!(
     op::Union{typeof(+),typeof(-)},
     A::Matrix,
     B::LinearAlgebra.UniformScaling,
 )
     n = LinearAlgebra.checksquare(A)
     for i = 1:n
-        A[i, i] = operate!(op, A[i, i], B)
+        A[i, i] = operate!!(op, A[i, i], B)
     end
     return A
 end
 
-function mutable_operate!(
+function operate!(
     op::AddSubMul,
     A::Matrix,
     B::Scaling,
     C::Scaling,
     D::Vararg{Scaling,N},
 ) where {N}
-    return mutable_operate!(add_sub_op(op), A, *(B, C, D...))
+    return operate!(add_sub_op(op), A, *(B, C, D...))
 end
 
 mul_rhs(::typeof(+)) = add_mul
@@ -77,26 +77,26 @@ function _check_dims(A, B)
     end
 end
 
-function mutable_operate!(
+function operate!(
     op::Union{typeof(+),typeof(-)},
     A::Array,
     B::AbstractArray,
 )
     _check_dims(A, B)
-    return mutable_broadcast!(op, A, B)
+    return broadcast!(op, A, B)
 end
 
 # We call `scaling_to_number` as `UniformScaling` do not support broadcasting
-function mutable_operate!(
+function operate!(
     op::AddSubMul,
     A::Array,
     B::AbstractArray,
     α::Vararg{Scaling,M},
 ) where {M}
     _check_dims(A, B)
-    return mutable_broadcast!(op, A, B, scaling_to_number.(α)...)
+    return broadcast!(op, A, B, scaling_to_number.(α)...)
 end
-function mutable_operate!(
+function operate!(
     op::AddSubMul,
     A::Array,
     α::Scaling,
@@ -104,7 +104,7 @@ function mutable_operate!(
     β::Vararg{Scaling,M},
 ) where {M}
     _check_dims(A, B)
-    return mutable_broadcast!(
+    return broadcast!(
         op,
         A,
         scaling_to_number(α),
@@ -112,7 +112,7 @@ function mutable_operate!(
         scaling_to_number.(β)...,
     )
 end
-function mutable_operate!(
+function operate!(
     op::AddSubMul,
     A::Array,
     α1::Scaling,
@@ -121,7 +121,7 @@ function mutable_operate!(
     β::Vararg{Scaling,M},
 ) where {M}
     _check_dims(A, B)
-    return mutable_broadcast!(
+    return broadcast!(
         op,
         A,
         scaling_to_number(α1),
@@ -133,13 +133,13 @@ end
 
 # Fallback, we may be able to be more efficient in more cases by adding more
 # specialized methods.
-function mutable_operate!(op::AddSubMul, A::Array, x, y)
-    return mutable_operate!(op, A, x * y)
+function operate!(op::AddSubMul, A::Array, x, y)
+    return operate!(op, A, x * y)
 end
 
-function mutable_operate!(op::AddSubMul, A::Array, x, y, args::Vararg{Any,N}) where {N}
+function operate!(op::AddSubMul, A::Array, x, y, args::Vararg{Any,N}) where {N}
     @assert N > 0
-    return mutable_operate!(op, A, x, *(y, args...))
+    return operate!(op, A, x, *(y, args...))
 end
 
 # Product
@@ -269,7 +269,7 @@ function _add_mul_array(buffer, C::Vector, A::AbstractMatrix, B::AbstractVector)
             aoffs = (k - 1) * Astride
[…]
