+++ src/shortcuts.jl
@@ -1,44 +1,44 @@
 """
-    add_to!(a, b, c)
+    add_to!!(a, b, c)
 
 Return the sum of `b` and `c`, possibly modifying `a`.
 """
-add_to!(output, args::Vararg{Any,N}) where {N} = operate_to!(output, +, args...)
+add_to!!(output, args::Vararg{Any,N}) where {N} = operate_to!!(output, +, args...)
 
 """
-    add!(a, b, ...)
+    add!!(a, b, ...)
 
 Return the sum of `a`, `b`, ..., possibly modifying `a`.
 """
-add!(args::Vararg{Any,N}) where {N} = operate!(+, args...)
+add!!(args::Vararg{Any,N}) where {N} = operate!!(+, args...)
 
 """
-    sub_to!(output, a, b)
+    sub_to!!(output, a, b)
 
 Return the `a - b`, possibly modifying `output`.
 """
-sub_to!(output, a, b) = operate_to!(output, -, a, b)
+sub_to!!(output, a, b) = operate_to!!(output, -, a, b)
 
 """
-    sub!(a, b)
+    sub!!(a, b)
 
 Return `a - b`, possibly modifying `a`.
 """
-sub!(a, b) = operate!(-, a, b)
+sub!!(a, b) = operate!!(-, a, b)
 
 """
-    mul_to!(a, b, c, ...)
+    mul_to!!(a, b, c, ...)
 
 Return the product of `b`, `c`, ..., possibly modifying `a`.
 """
-mul_to!(output, args::Vararg{Any,N}) where {N} = operate_to!(output, *, args...)
+mul_to!!(output, args::Vararg{Any,N}) where {N} = operate_to!!(output, *, args...)
 
 """
-    mul!(a, b, ...)
+    mul!!(a, b, ...)
 
 Return the product of `a`, `b`, ..., possibly modifying `a`.
 """
-mul!(args::Vararg{Any,N}) where {N} = operate!(*, args...)
+mul!!(args::Vararg{Any,N}) where {N} = operate!!(*, args...)
 
 """
     mul(a, b, ...)
@@ -48,107 +48,107 @@ Shortcut for `operate(*, a, b, ...)`, see [`operate`](@ref).
 mul(args::Vararg{Any,N}) where {N} = operate(*, args...)
 
 """
-    gcd_to!(a, b, c, ...)
+    gcd_to!!(a, b, c, ...)
 
 Return the `gcd` of `b`, `c`, ..., possibly modifying `a`.
 """
-gcd_to!(output, args::Vararg{Any,N}) where {N} = operate_to!(output, gcd, args...)
+gcd_to!!(output, args::Vararg{Any,N}) where {N} = operate_to!!(output, gcd, args...)
 
 """
-    gcd!(a, b, ...)
+    gcd!!(a, b, ...)
 
 Return the `gcd` of `a`, `b`, ..., possibly modifying `a`.
 """
-gcd!(args::Vararg{Any,N}) where {N} = operate!(gcd, args...)
+gcd!!(args::Vararg{Any,N}) where {N} = operate!!(gcd, args...)
 
 """
-    lcm_to!(a, b, c, ...)
+    lcm_to!!(a, b, c, ...)
 
 Return the `lcm` of `b`, `c`, ..., possibly modifying `a`.
 """
-lcm_to!(output, args::Vararg{Any,N}) where {N} = operate_to!(output, lcm, args...)
+lcm_to!!(output, args::Vararg{Any,N}) where {N} = operate_to!!(output, lcm, args...)
 
 """
-    lcm!(a, b, ...)
+    lcm!!(a, b, ...)
 
 Return the `lcm` of `a`, `b`, ..., possibly modifying `a`.
 """
-lcm!(args::Vararg{Any,N}) where {N} = operate!(lcm, args...)
[…]
