### 2021-10-20 b3d8ea8
Fix hygiene of promote_operation (#123)
* Fix hygiene of promote_operation

* Add test
