+++ paper/paper.tex
@@ -23,13 +23,13 @@ This talk presents an interface that allows algorithms to exploit mutability in 
 
 \section{Introduction}
 
-Julia enables generic algorithms that work with arbitrary number types, as long as the types implement the needed operations
-such as \texttt{+}, \texttt{*}, \texttt{-}, \texttt{zero}, \texttt{one}, ...
+Julia enables developers to write generic algorithms that work with arbitrary number types, as long as the types implement the needed operations
+such as \lstinline|+|, \lstinline|*|, \lstinline|-|, \lstinline|zero|, \lstinline|one|, ...
 The implementations of these arithmetic operations in Julia do not modify their arguments. Instead, they return a new instance of the type as the result.
-However, in many situations, a variable represents an accumulator that can be modified to contain the result, e.g.,
+However, in many situations, a variable represents an accumulator that can be mutated\footnote{\label{foot:mutate}In this paper, the terminology ``mutate $x$ to $y$'' means modifying $x$ in-place in such a way that its value after the modification is equal to $y$} to the result, e.g.,
 when summing the elements of an array or when implementing array multiplication.
 Moreover, for types that support mutation, mutating the value may have a significant performance benefit over creating a new instance.
-Examples of types that implement arithmetic operations and support mutation are \texttt{Array}s, multiple precision numbers, JuMP~\cite{dunning2017jump} expressions, MathOptInterface (MOI)~\cite{legat2021mathoptinterface} functions, and polynomials (univariate~\cite{verzani2021polynomials} or multivariate~\cite{legat2021multivariatepolynomials}).
+Examples of types that implement arithmetic operations and support mutation include \lstinline|Array|s, multiple precision numbers, JuMP~\cite{dunning2017jump} expressions, MathOptInterface (MOI)~\cite{legat2021mathoptinterface} functions, and polynomials (univariate~\cite{verzani2021polynomials} or multivariate~\cite{legat2021multivariatepolynomials}).
 
 This paper introduces an interface called \ma{}.
 It allows mutable types to implement an arithmetic exploiting their mutability, and for algorithms to
@@ -41,7 +41,7 @@ Moreover, it provides the following additional features:
     it re-implements part of the Julia standard library on top of the API to allow mutable types to use a more efficient version than the default one.
   \item
     \label{item:rewrite}
-    it defines a \texttt{@rewrite} macro that rewrites an expression using the standard operations (e.g \texttt{+}, \texttt{*}, ...) into an expression that exploits the mutability of the intermediate values created when evaluating the expression.
+    it defines a \lstinline|@rewrite| macro that rewrites an expression using the standard operations (e.g \lstinline|+|, \lstinline|*|, ...) into an expression that exploits the mutability of the intermediate values created when evaluating the expression.
 \end{enumerate}
 
 JuMP~\cite{dunning2017jump} used to have its own API for mutable operations on JuMP expressions and
@@ -58,7 +58,7 @@ The section is organized into four subsections that describe the need of four ke
 \subsection{May mutate}
 \label{sec:may_mutate}
 Consider the task of summing the elements of a vector.
-By default, Julia \texttt{sum} function will compute with a code equivalent to the following:
+By default, Julia \lstinline|sum| function will compute the sum with a code equivalent to the following:
 \begin{lstlisting}[language = Julia]
 function sum(x::Vector)
     acc = zero(eltype(x))
[…]
