### 2021-11-03 #93 @ksil
"This examples encapsulates" on page 1 should be "This example encapsulates", and "is a the core" should be "is at the core" on page 3. Otherwise, with those changes, everything looks good to me @ranjanan
