### 2021-11-03 #93 @blegat
Thanks your for noticing this and sorry I missed these in your comments, I just fixed these two typos.
