+++ src/bigfloat.jl
@@ -1,5 +1,11 @@
 mutability(::Type{BigFloat}) = IsMutable()
-mutable_copy(x::BigFloat) = deepcopy(x)
+# copied from `deepcopy_internal` implementation in Julia:
+# https://github.com/JuliaLang/julia/blob/7d41d1eb610cad490cbaece8887f9bbd2a775021/base/mpfr.jl#L1041-L1050
+function mutable_copy(x::BigFloat)
+    d = x._d
+    d′ = GC.@preserve d unsafe_string(pointer(d), sizeof(d)) # creates a definitely-new String
+    return Base.MPFR._BigFloat(x.prec, x.sign, x.exp, d′)
+end
 
 @static if VERSION >= v"1.1.0-DEV.683"
     const _MPFRRoundingMode = Base.MPFR.MPFRRoundingMode
