### 2021-11-11 be5f895
Implement operate for /
