+++ test/range.jl
@@ -6,7 +6,7 @@ function mutating_step_range_test(::Type{T}) where T
     r = MA.MutatingStepRange(T(2), T(3), T(9))
     expected = MA.mutability(T) isa MA.IsMutable ? 8 * ones(T, 3) : T[2, 5, 8]
     @test collect(r) == expected
-    @test reduce(MA.add!, r) == T(15)
+    @test reduce(MA.add!!, r) == T(15)
 end
 
 @testset "MutatingStepRange" begin
