### 2021-11-11 419c77f
Implement operate for / (#125)
