### 2021-12-01 61dbe44
Fix ambiguity with new mul! methods in Julia v1.7
