### 2021-12-01 c23ff2d
Workaround for mat_mat_scalar
