### 2021-12-26 f0e61aa
Workaround for mat_mat_scalar (#126)
