### 2022-01-17 c5f4f6d
README.md: update latest -> dev in doc URLs (#131)
This was changed deprecated in Documenter.jl 0.20.0 and dropped in 0.24.0
