### 2022-02-04 be8dbb4
Add JuliaFormatter and add CI check
