+++ test/matmul.jl
@@ -13,8 +13,10 @@ function dot_test(x, y)
     @test MA.operate(LinearAlgebra.dot, y, x) == LinearAlgebra.dot(y, x)
     @test MA.operate(*, x', y) == x' * y
     @test MA.operate(*, y', x) == y' * x
-    @test MA.operate(*, LinearAlgebra.transpose(x), y) == LinearAlgebra.transpose(x) * y
-    @test MA.operate(*, LinearAlgebra.transpose(y), x) == LinearAlgebra.transpose(y) * x
+    @test MA.operate(*, LinearAlgebra.transpose(x), y) ==
+          LinearAlgebra.transpose(x) * y
+    @test MA.operate(*, LinearAlgebra.transpose(y), x) ==
+          LinearAlgebra.transpose(y) * x
 end
 
 @testset "dot" begin
@@ -41,12 +43,17 @@ end
 end
 
 @testset "convert" begin
-    @test MA.scaling_convert(LinearAlgebra.UniformScaling{Int}, LinearAlgebra.I) isa
-          LinearAlgebra.UniformScaling
+    @test MA.scaling_convert(
+        LinearAlgebra.UniformScaling{Int},
+        LinearAlgebra.I,
+    ) isa LinearAlgebra.UniformScaling
     @test MA.scaling_convert(Int, LinearAlgebra.I) === 1
     @test MA.scaling_convert(Int, 1) === 1
-    @test MA.operate(convert, LinearAlgebra.UniformScaling{Int}, LinearAlgebra.I) isa
-          LinearAlgebra.UniformScaling
+    @test MA.operate(
+        convert,
+        LinearAlgebra.UniformScaling{Int},
+        LinearAlgebra.I,
+    ) isa LinearAlgebra.UniformScaling
     @test MA.operate(convert, Int, LinearAlgebra.I) === 1
     @test MA.operate(convert, Int, 1) === 1
 end
@@ -64,13 +71,17 @@ struct NoProdMutable <: MA.AbstractMutable end
     # Hack for making the test work on old Julia versions
     Base.:*(::NoProdMutable, ::NoProdMutable) = error(EXPECTED_ERROR)
 else
-    function MA.promote_operation(::typeof(*), ::Type{NoProdMutable}, ::Type{NoProdMutable})
+    function MA.promote_operation(
+        ::typeof(*),
+        ::Type{NoProdMutable},
+        ::Type{NoProdMutable},
+    )
         return Int # Dummy result just to test error message
     end
 end
 
 function unsupported_product()
-    A = [NoProdMutable() for i = 1:2, j = 1:2]
+    A = [NoProdMutable() for i in 1:2, j in 1:2]
     err = ErrorException(EXPECTED_ERROR)
     @test_throws err A * A
 end
@@ -104,7 +115,9 @@ end
         )
         @test_throws err MA.operate(*, x', y)
         @test_throws err MA.operate(*, LinearAlgebra.transpose(x), y)
-        err = DimensionMismatch("matrix A has dimensions (2,2), vector B has length 1")
+        err = DimensionMismatch(
+            "matrix A has dimensions (2,2), vector B has length 1",
+        )
         @test_throws err MA.operate(*, x', B)
         a = zeros(0)
         @test iszero(@inferred MA.operate(LinearAlgebra.dot, a, a))
@@ -139,7 +152,11 @@ end
         @test MA.mul(A, x) == BigInt[3; 3; 3]
         @test MA.mul_to!!(y, A, x) == BigInt[3; 3; 3] && y == BigInt[3; 3; 3]
         @test_throws DimensionMismatch MA.mul(BigInt[1 1; 1 1], BigInt[])
[…]
