+++ src/Test/config.jl
@@ -6,14 +6,21 @@ macro test_suite(setname, subsets = false)
     else
         runtest = :(f(args...))
     end
-    esc(:(function $testname(args...; exclude::Vector{String} = String[]) where {T}
-        for (name, f) in $testdict
-            if name in exclude
-                continue
+    return esc(
+        :(
+            function $testname(
+                args...;
+                exclude::Vector{String} = String[],
+            ) where {T}
+                for (name, f) in $testdict
+                    if name in exclude
+                        continue
+                    end
+                    @testset "$name" begin
+                        $runtest
+                    end
+                end
             end
-            @testset "$name" begin
-                $runtest
-            end
-        end
-    end))
+        ),
+    )
 end

[…]
