### 2022-02-04 2eda0df
Add JuliaFormatter and add CI check (#133)
