+++ src/reduce.jl
@@ -7,7 +7,7 @@ function _same_length(a, b, c::Vararg{Any,N}) where {N}
             ),
         )
     end
-    _same_length(b, c...)
+    return _same_length(b, c...)
 end
 reduce_op(op::AddSubMul) = add_sub_op(op)
 reduce_op(op::typeof(add_dot)) = +
@@ -15,28 +15,18 @@ neutral_element(::typeof(+), T::Type) = zero(T)
 map_op(::AddSubMul) = *
 map_op(::typeof(add_dot)) = LinearAlgebra.dot
 function promote_map_reduce(op::Function, args::Vararg{Any,N}) where {N}
-    T = promote_operation(
-        op,
-        promote_operation(map_op(op), args...),
-        args...,
-    )
+    return T =
+        promote_operation(op, promote_operation(map_op(op), args...), args...)
 end
 
-function fused_map_reduce(
-    op::F,
-    args::Vararg{Any,N},
-) where {F<:Function,N}
+function fused_map_reduce(op::F, args::Vararg{Any,N}) where {F<:Function,N}
     _same_length(args...)
     T = promote_map_reduce(op, eltype.(args)...)
     accumulator = neutral_element(reduce_op(op), T)
     buffer = buffer_for(op, T, eltype.(args)...)
     for I in zip(eachindex.(args)...)
-        accumulator = buffered_operate!!(
-            buffer,
-            op,
-            accumulator,
-            getindex.(args, I)...,
-        )
+        accumulator =
+            buffered_operate!!(buffer, op, accumulator, getindex.(args, I)...)
     end
     return accumulator
 end
