+++ test/dummy.jl
@@ -15,22 +15,42 @@ Base.broadcastable(x::DummyBigInt) = Ref(x)
 # Julia v1.6+.
 Base.ndims(::Union{Type{DummyBigInt},DummyBigInt}) = 0
 
-Base.promote_rule(::Type{DummyBigInt}, ::Type{<:Union{Integer,UniformScaling{<:Integer}}}) =
-    DummyBigInt
+function Base.promote_rule(
+    ::Type{DummyBigInt},
+    ::Type{<:Union{Integer,UniformScaling{<:Integer}}},
+)
+    return DummyBigInt
+end
 # `copy` on BigInt returns the same instance anyway
 Base.copy(x::DummyBigInt) = x
 MA.mutable_copy(x::DummyBigInt) = DummyBigInt(MA.mutable_copy(x.data))
 LinearAlgebra.symmetric_type(::Type{DummyBigInt}) = DummyBigInt
 LinearAlgebra.symmetric(x::DummyBigInt, ::Symbol) = x
 LinearAlgebra.dot(x::DummyBigInt, y::DummyBigInt) = x * y
-LinearAlgebra.dot(x::DummyBigInt, y::Union{Integer,UniformScaling{<:Integer}}) = x * y
-LinearAlgebra.dot(x::Union{Integer,UniformScaling{<:Integer}}, y::DummyBigInt) = x * y
+function LinearAlgebra.dot(
+    x::DummyBigInt,
+    y::Union{Integer,UniformScaling{<:Integer}},
+)
+    return x * y
+end
+function LinearAlgebra.dot(
+    x::Union{Integer,UniformScaling{<:Integer}},
+    y::DummyBigInt,
+)
+    return x * y
+end
 LinearAlgebra.transpose(x::DummyBigInt) = x
 LinearAlgebra.adjoint(x::DummyBigInt) = x
 MA.mutability(::Type{DummyBigInt}) = MA.IsMutable()
 MA.promote_operation(::typeof(zero), ::Type{DummyBigInt}) = DummyBigInt
 MA.promote_operation(::typeof(one), ::Type{DummyBigInt}) = DummyBigInt
-MA.promote_operation(::typeof(+), ::Type{DummyBigInt}, ::Type{DummyBigInt}) = DummyBigInt
+function MA.promote_operation(
+    ::typeof(+),
+    ::Type{DummyBigInt},
+    ::Type{DummyBigInt},
+)
+    return DummyBigInt
+end
 _data(x) = x
 _data(x::DummyBigInt) = x.data
 MA.scaling(x::DummyBigInt) = x
@@ -43,7 +63,10 @@ function MA.operate_to!(
     return DummyBigInt(MA.operate_to!(x.data, op, _data.(args)...))
 end
 
-function MA.buffer_for(::MA.AddSubMul, args::Vararg{Type{DummyBigInt},N}) where {N}
+function MA.buffer_for(
+    ::MA.AddSubMul,
+    args::Vararg{Type{DummyBigInt},N},
+) where {N}
     return DummyBigInt(BigInt())
 end
 _undummy(x) = x
@@ -54,7 +77,13 @@ function MA.buffered_operate_to!(
     op::Function,
     args::Vararg{Union{MA.Scaling,DummyBigInt},N},
 ) where {N}
-    MA.buffered_operate_to!(buffer.data, output.data, output, op, _undummy.(args)...)
+    MA.buffered_operate_to!(
+        buffer.data,
+        output.data,
+        output,
+        op,
+        _undummy.(args)...,
+    )
     return output
 end
 function MA.buffered_operate!(
@@ -80,7 +109,11 @@ end
[…]
