+++ src/implementations/BigFloat.jl
@@ -1,5 +1,6 @@
 mutability(::Type{BigFloat}) = IsMutable()
-# copied from `deepcopy_internal` implementation in Julia:
+
+# Copied from `deepcopy_internal` implementation in Julia:
 # https://github.com/JuliaLang/julia/blob/7d41d1eb610cad490cbaece8887f9bbd2a775021/base/mpfr.jl#L1041-L1050
 function mutable_copy(x::BigFloat)
     d = x._d
@@ -14,7 +15,9 @@ else
 end
 
 # zero
+
 promote_operation(::typeof(zero), ::Type{BigFloat}) = BigFloat
+
 function _set_si!(x::BigFloat, value)
     ccall(
         (:mpfr_set_si, :libmpfr),
@@ -29,11 +32,15 @@ end
 operate!(::typeof(zero), x::BigFloat) = _set_si!(x, 0)
 
 # one
+
 promote_operation(::typeof(one), ::Type{BigFloat}) = BigFloat
+
 operate!(::typeof(one), x::BigFloat) = _set_si!(x, 1)
 
 # +
+
 promote_operation(::typeof(+), ::Vararg{Type{BigFloat},N}) where {N} = BigFloat
+
 function operate_to!(output::BigFloat, ::typeof(+), a::BigFloat, b::BigFloat)
     ccall(
         (:mpfr_add, :libmpfr),
@@ -46,12 +53,11 @@ function operate_to!(output::BigFloat, ::typeof(+), a::BigFloat, b::BigFloat)
     )
     return output
 end
-#function operate_to!(output::BigFloat, op::typeof(+), a::BigFloat, b::LinearAlgebra.UniformScaling)
-#    return operate_to!(output, op, a, b.λ)
-#end
 
 # -
+
 promote_operation(::typeof(-), ::Vararg{Type{BigFloat},N}) where {N} = BigFloat
+
 function operate_to!(output::BigFloat, ::typeof(-), a::BigFloat, b::BigFloat)
     ccall(
         (:mpfr_sub, :libmpfr),
@@ -66,7 +72,9 @@ function operate_to!(output::BigFloat, ::typeof(-), a::BigFloat, b::BigFloat)
 end
 
 # *
+
 promote_operation(::typeof(*), ::Vararg{Type{BigFloat},N}) where {N} = BigFloat
+
 function operate_to!(output::BigFloat, ::typeof(*), a::BigFloat, b::BigFloat)
     ccall(
         (:mpfr_mul, :libmpfr),
@@ -90,13 +98,16 @@ function operate_to!(
     operate_to!(output, op, a, b)
     return operate!(op, output, c...)
 end
+
 function operate!(op::Function, x::BigFloat, args::Vararg{Any,N}) where {N}
     return operate_to!(x, op, x, args...)
 end
 
 # add_mul and sub_mul
+
 # Buffer to hold the product
 buffer_for(::AddSubMul, args::Vararg{Type{BigFloat},N}) where {N} = BigFloat()
+
 function operate_to!(
     output::BigFloat,
     op::AddSubMul,
@@ -120,6 +131,7 @@ function buffered_operate_to!(
     operate_to!(buffer, *, x, y, args...)
     return operate_to!(output, add_sub_op(op), a, buffer)
 end
+
 function buffered_operate!(
     buffer::BigFloat,
     op::AddSubMul,
@@ -138,6 +150,7 @@ function operate_to!(
 ) where {N}
     return operate_to!(output, op, _scaling_to_bigfloat.(args)...)
 end
+
 function operate_to!(
     output::BigFloat,
     op::AddSubMul,
@@ -155,6 +168,7 @@ function operate_to!(
         _scaling_to_bigfloat.(args)...,
[…]
