### 2022-02-04 b08585f
More refactoring
