+++ src/implementations/BigFloat.jl
@@ -1,3 +1,6 @@
+# This file contains methods to implement the MutableArithmetics API for
+# Base.BigFloat.
+
 mutability(::Type{BigFloat}) = IsMutable()
 
 # Copied from `deepcopy_internal` implementation in Julia:
