+++ src/implementations/Rational.jl
@@ -1,10 +1,16 @@
-# `Rational` is a `struct`, not a `mutable struct` so if `T` is not
+# This file contains methods to implement the MutableArithmetics API for
+# Base.Rational{T}.
+
+# `Rational` is a `struct`, not a `mutable struct`, so if `T` is not
 # mutable, we cannot mutate the rational.
 mutability(::Type{Rational{T}}) where {T} = mutability(T)
+
 mutable_copy(x::Rational) = Rational(mutable_copy(x.num), mutable_copy(x.den))
 
 # zero
+
 promote_operation(::typeof(zero), ::Type{Rational{T}}) where {T} = Rational{T}
+
 function operate!(::typeof(zero), x::Rational)
     operate!(zero, x.num)
     operate!(one, x.den)
@@ -12,7 +18,9 @@ function operate!(::typeof(zero), x::Rational)
 end
 
 # one
+
 promote_operation(::typeof(one), ::Type{Rational{T}}) where {T} = Rational{T}
+
 function operate!(::typeof(one), x::Rational)
     operate!(one, x.num)
     operate!(one, x.den)
@@ -20,6 +28,7 @@ function operate!(::typeof(one), x::Rational)
 end
 
 # +
+
 function promote_operation(
     ::typeof(+),
     ::Type{Rational{S}},
@@ -27,9 +36,10 @@ function promote_operation(
 ) where {S,T}
     return Rational{promote_sum_mul(S, T)}
 end
+
 function operate_to!(output::Rational, ::typeof(+), x::Rational, y::Rational)
     xd, yd = Base.divgcd(promote(x.den, y.den)...)
-    # TODO Use `checked_mul` and `checked_add` like in Base
+    # TODO: Use `checked_mul` and `checked_add` like in Base
     operate_to!(output.num, *, x.num, yd)
     operate!(add_mul, output.num, y.num, xd)
     operate_to!(output.den, *, x.den, yd)
@@ -37,6 +47,7 @@ function operate_to!(output::Rational, ::typeof(+), x::Rational, y::Rational)
 end
 
 # -
+
 function promote_operation(
     ::typeof(-),
     ::Type{Rational{S}},
@@ -44,9 +55,10 @@ function promote_operation(
 ) where {S,T}
     return Rational{promote_sum_mul(S, T)}
 end
+
 function operate_to!(output::Rational, ::typeof(-), x::Rational, y::Rational)
     xd, yd = Base.divgcd(promote(x.den, y.den)...)
-    # TODO Use `checked_mul` and `checked_sub` like in Base
+    # TODO: Use `checked_mul` and `checked_sub` like in Base
     operate_to!(output.num, *, x.num, yd)
     operate!(sub_mul, output.num, y.num, xd)
     operate_to!(output.den, *, x.den, yd)
@@ -54,6 +66,7 @@ function operate_to!(output::Rational, ::typeof(-), x::Rational, y::Rational)
 end
 
 # *
+
 function promote_operation(
     ::typeof(*),
     ::Type{Rational{S}},
@@ -61,6 +74,7 @@ function promote_operation(
 ) where {S,T}
     return Rational{promote_operation(*, S, T)}
 end
+
 function operate_to!(output::Rational, ::typeof(*), x::Rational, y::Rational)
     xn, yd = Base.divgcd(promote(x.num, y.den)...)
     xd, yn = Base.divgcd(promote(x.den, y.num)...)
@@ -70,6 +84,7 @@ function operate_to!(output::Rational, ::typeof(*), x::Rational, y::Rational)
 end
 
 # gcd
+
 function promote_operation(
     ::Union{typeof(gcd),typeof(lcm)},
     ::Type{Rational{S}},
@@ -77,11 +92,13 @@ function promote_operation(
[…]
