### 2022-02-04 64c02d0
Style improvements in dispatch.jl
