+++ src/dispatch.jl
@@ -1,7 +1,8 @@
-# TODO: Intercepting "externally owned" method calls by dispatching on type parameters
-# (rather than outermost wrapper type) is generally bad practice, but refactoring this code
-# to use a different mechanism would be a lot of work. In the future, this interception code
-# would be more easily/robustly replaced by using a tool like
+# TODO: this file contains a large number of method specializations to intercept
+# "externally owned" method calls by dispatching on type parameters (rather than
+# outermost wrapper type). This is generally bad practice, but refactoring this
+# code to use a different mechanism would be a lot of work. In the future, this
+# interception code would be more easily/robustly replaced by using a tool like
 # https://github.com/jrevels/Cassette.jl.
 
 abstract type AbstractMutable end
@@ -10,18 +11,21 @@ function Base.sum(a::AbstractArray{<:AbstractMutable})
     return operate(sum, a)
 end
 
-# When doing `x'y` where the elements of `x` and/or `y` are arrays,
-# redirecting to `dot(x, y)` is not equivalent to `x'y` as it will call
-# dot recursively on the elements of `x` and `y`:
-# See https://github.com/JuliaLang/julia/issues/35174
-# For this reason, a `_dot_nonrecursive` function was added that does
-# not recursively call `dot`:
+# When doing `x'y` where the elements of `x` and/or `y` are arrays, redirecting
+# to `dot(x, y)` is not equivalent to `x'y` as it will call dot recursively on
+# the elements of `x` and `y`. See
+# https://github.com/JuliaLang/julia/issues/35174
+# For this reason, a `_dot_nonrecursive` function was added that does not
+# recursively call `dot`:
 # https://github.com/JuliaLang/julia/commit/eae3216416453b53631afa6c803591cf2c5ae5b3
-# However, it does not exploit mutability and returns `zero(eltype(lhs)) * zero(eltype(rhs))`
-# in case the arrays are empty which creates type unstability for some types for which this
-# type is not invariant under addition.
-# TODO LinearAlgebra should have a documented function so that we don't
-#      have to overload an internal function
+#
+# However, it does not exploit mutability and returns
+# `zero(eltype(lhs)) * zero(eltype(rhs))` in case the arrays are empty which
+# creates type instability for some types for which this type is not invariant
+# under addition.
+
+# TODO: LinearAlgebra should have a documented function so that we don't have to
+# overload an internal function
 if VERSION >= v"1.5.0-rc1.23"
     function LinearAlgebra._dot_nonrecursive(
         lhs::AbstractArray{<:AbstractMutable},
@@ -49,12 +53,14 @@ function LinearAlgebra.dot(
 )
     return operate(LinearAlgebra.dot, lhs, rhs)
 end
+
 function LinearAlgebra.dot(
     lhs::AbstractArray,
     rhs::AbstractArray{<:AbstractMutable},
 )
     return operate(LinearAlgebra.dot, lhs, rhs)
 end
+
 function LinearAlgebra.dot(
[…]
