### 2022-02-04 f91300b
Style improvements in reduce.jl
