### 2022-02-04 5f5be9c
Style improvements for rewrite.jl
