+++ src/MutableArithmetics.jl
@@ -9,6 +9,8 @@ module MutableArithmetics
 import LinearAlgebra
 import SparseArrays
 
+using Base.Meta  # Used in rewrite.jl
+
 # Performance note:
 # We use `Vararg` instead of splatting `...` as using `where N` forces Julia to
 # specialize in the number of arguments `N`. Otherwise, we get allocations and
