### 2022-02-04 9cd15bc
Style improvements in reduce.jl (#136)
* Style improvements in reduce.jl

* Update reduce.jl
