### 2022-02-06 1c9f375
add unary minus on BigInt
