+++ src/interface.jl
@@ -172,9 +172,7 @@ function operate(
 ) where {N}
     return op(x, y, args...)
 end
-
-operate(op::Union{typeof(-),typeof(/)}, x, y) where {N} = op(x, y)
-
+operate(op::Union{typeof(-),typeof(/),typeof(//)}, x, y) = op(x, y)
 operate(::typeof(convert), ::Type{T}, x) where {T} = convert(T, x)
 
 operate(::typeof(convert), ::Type{T}, x::T) where {T} = copy_if_mutable(x)
