### 2022-02-06 6833e46
add more complete rational arithmetic
