+++ src/implementations/Rational.jl
@@ -28,13 +28,28 @@ function operate!(::typeof(one), x::Rational)
 end
 
 # +
-
 function promote_operation(
-    ::typeof(+),
+    ::Union{typeof(+),typeof(-)},
     ::Type{Rational{S}},
-    ::Type{Rational{T}},
+    ::Type{Rational{T}}
 ) where {S,T}
     return Rational{promote_sum_mul(S, T)}
+end
+
+function promote_operation(
+    op::Union{typeof(+),typeof(-)},
+    ::Type{Rational{S}},
+    ::Type{I}
+) where {S,I<:Integer}
+    return promote_operation(op, Rational{S}, Rational{I})
+end
+
+function promote_operation(
+    op::Union{typeof(+),typeof(-)},
+    ::Type{I},
+    ::Type{Rational{S}}
+) where {S,I<:Integer}
+    return promote_operation(op, Rational{S}, Rational{I})
 end
 
 function operate_to!(output::Rational, ::typeof(+), x::Rational, y::Rational)
@@ -46,22 +61,48 @@ function operate_to!(output::Rational, ::typeof(+), x::Rational, y::Rational)
     return output
 end
 
-# -
-
-function promote_operation(
-    ::typeof(-),
-    ::Type{Rational{S}},
-    ::Type{Rational{T}},
-) where {S,T}
-    return Rational{promote_sum_mul(S, T)}
+function operate_to!(output::Rational, ::typeof(+), x::Rational, y::Integer)
+    # TODO Use `checked_mul` and `checked_add` like in Base
+    operate_to!(output.num, *, x.den, y)
+    operate!(+, output.num, x.num)
+    operate_to!(output.den, *, x.den, oftype(x.den, 1))
+    return output
 end
 
+function operate_to!(output::Rational, ::typeof(+), y::Integer, x::Rational)
+    return operate_to!(output, +, x, y)
+end
+
+# unary -
+function operate_to!(output::Rational, ::typeof(-), x::Rational)
+    operate_to!(output.num, -, x.num)
+    operate_to!(output.den, copy, x.den)
+    return output
+end
+
+# binary -
 function operate_to!(output::Rational, ::typeof(-), x::Rational, y::Rational)
     xd, yd = Base.divgcd(promote(x.den, y.den)...)
     # TODO: Use `checked_mul` and `checked_sub` like in Base
     operate_to!(output.num, *, x.num, yd)
     operate!(sub_mul, output.num, y.num, xd)
     operate_to!(output.den, *, x.den, yd)
+    return output
+end
+
+function operate_to!(output::Rational, ::typeof(-), x::Rational, y::Integer)
+    # TODO Use `checked_mul` and `checked_sub` like in Base
+    operate_to!(output.num, *, x.den, y)
+    operate!(-, output.num)
+    operate!(+, output.num, x.num)
+    operate_to!(output.den, copy, x.den)
+    return output
+end
+
+function operate_to!(output::Rational, ::typeof(-), y::Integer, x::Rational)
+    # TODO Use `checked_mul` and `checked_sub` like in Base
+    operate_to!(output, -, x, y)
+    operate_to!(output, -, output)
     return output
 end
 
@@ -75,11 +116,69 @@ function promote_operation(
     return Rational{promote_operation(*, S, T)}
 end
 
+function promote_operation(
+    ::typeof(*),
+    ::Type{Rational{S}},
+    ::Type{I}
+) where {S,I<:Integer}
+    return promote_operation(*, Rational{S}, Rational{I})
+end
+
+function promote_operation(
+    ::typeof(*),
+    ::Type{I},
+    ::Type{Rational{S}}
+) where {S,I<:Integer}
[…]
