### 2022-02-06 d4f8d2e
use Base.promote_op in  promote_operation_fallback
