### 2022-02-07 20de975
Fix for StaticArrays in multiplication
