### 2022-02-07 8b873df
Clarify docstring of operate_to
