+++ README.md
@@ -8,66 +8,63 @@
 [![version][version-img]][version-url]
 [![pkgeval][pkgeval-img]][pkgeval-url]
 
-**MutableArithmetics** (MA for short) is a [Julia](http://julialang.org) package which allows:
-*   for mutable types to implement mutable arithmetics;
-*   for algorithms that could exploit mutable arithmetics to exploit them while still being completely generic.
+**MutableArithmetics** (MA for short) is a [Julia](http://julialang.org) package
+which allows:
+ * for mutable types to implement mutable arithmetics
+ * for algorithms that could exploit mutable arithmetics to exploit them while
+   still being completely generic.
 
 While in some cases, similar features have been included in packages
-idiosyncratically, the goal of this package is to provide a generic interface to
-allow anyone to make use of mutability when desired.
+idiosyncratically, the goal of MutableArithmetics is to provide a generic
+interface to allow anyone to make use of mutability when desired.
 
-The package allows a given type to declare itself mutable through the
-`MA.mutability` trait.
-Then the user can use the `MA.operate!!` function to write generic code
+The package allows a type to declare itself mutable through the `MA.mutability`
+trait. Then the user can use the `MA.operate!!` function to write generic code
 that works for arbitrary type while exploiting mutability of the type
 if possible. More precisely:
+ * The `MA.operate!!(op::Function, x, args...)` redirects to `op(x, args...)`
+   if `x` is not mutable or if the result of the operation cannot be stored in
+   `x`. Otherwise, it redirects to `MA.operate!(op, x, args...)`.
+ * `MA.operate!(op::Function, x, args...)` stores the result of the operation in
+   `x`. It is a `MethodError` if `x` is not mutable or if the result of the
+   operation cannot be stored in `x`.
 
-* The `MA.operate!!(op::Function, x, args...)` redirects to `op(x, args...)`
-  if `x` is not mutable or if the result of the operation cannot be stored in `x`.
-  Otherwise, it redirects to `MA.operate!(op, x, args...)`.
-* `MA.operate!(op::Function, x, args...)` stores the result of the
-  operation in `x`. It is a `MethodError` if `x` is not mutable or if the
-  result of the operation cannot be stored in `x`.
+So from a generic code perspective, `MA.operate!!` can be used when the value of
+`x` is not used anywhere else. This allows the code to both work for mutable and
+for non-mutable type.
 
-So from a generic code, `MA.operate!!` can be used when the value of `x` is not
-used anywhere else to recycle it if possible. This allows the code to both
-work for mutable and for non-mutable type.
+When the type is known to be mutable, `MA.operate!` can be used to make sure the
+operation is done in-place. If it is not possible, the `MethodError` allows one
+to easily fix the issue while `MA.operate!!` would have silently fallen back to
+the non-mutating function.
 
-When the type is known to be mutable, `MA.operate!` can be used to make
-sure the operation is done in-place. If it is not possible, the `MethodError`
-allows to easily fix the issue while `MA.operate!!` would have silently fallen
-back to the non-mutating function.
+In conclusion, the distinction between `MA.operate!!` and `MA.operate!` covers
+all use case while having an universal convention accross all operations.
 
-In conclusion, the distinction between `MA.operate!!` and `MA.operate!`
-allows to cover all use case while having an universal convention accross all
-operations.
[…]
