+++ src/interface.jl
@@ -14,9 +14,11 @@ function promote_operation_fallback(
     # `zero` is not defined for `AbstractArray` so the fallback would fail with a cryptic MethodError.
     # We replace it by a more helpful error here.
     return error(
-        "`promote_operation($op, $x, $y)` not implemented yet, please report this.",
+        "`promote_operation($op, $x, $y)` not implemented yet, please report " *
+        "this.",
     )
 end
+
 function promote_operation_fallback(
     ::typeof(/),
     ::Type{S},
@@ -24,6 +26,7 @@ function promote_operation_fallback(
 ) where {S,T}
     return typeof(zero(S) / oneunit(T))
 end
+
 # Julia v1.0.x has trouble with inference with the `Vararg` method, see
 # https://travis-ci.org/jump-dev/JuMP.jl/jobs/617606373
 function promote_operation_fallback(
@@ -33,13 +36,16 @@ function promote_operation_fallback(
 ) where {F<:Function,S,T}
     return typeof(op(zero(S), zero(T)))
 end
+
 function promote_operation_fallback(
     op::F,
     args::Vararg{Type,N},
 ) where {F<:Function,N}
     return typeof(op(zero.(args)...))
 end
+
 promote_operation_fallback(::typeof(*), ::Type{T}) where {T} = T
+
 function promote_operation_fallback(
     ::typeof(*),
     ::Type{S},
@@ -50,10 +56,12 @@ function promote_operation_fallback(
     return promote_operation(*, promote_operation(*, S, T), U, args...)
 end
 
-# `Vararg` gives extra allocations on Julia v1.3, see https://travis-ci.com/jump-dev/MutableArithmetics.jl/jobs/260666164#L215-L238
+# `Vararg` gives extra allocations on Julia v1.3, see
+# https://travis-ci.com/jump-dev/MutableArithmetics.jl/jobs/260666164#L215-L238
 function promote_operation_fallback(op::AddSubMul, T::Type, x::Type, y::Type)
     return promote_operation(add_sub_op(op), T, promote_operation(*, x, y))
 end
+
 function promote_operation_fallback(
     op::AddSubMul,
     x::Type{<:AbstractArray},
@@ -61,6 +69,7 @@ function promote_operation_fallback(
 )
     return promote_operation(add_sub_op(op), x, y)
 end
+
 function promote_operation_fallback(
     op::Union{AddSubMul,typeof(add_dot)},
     T::Type,
@@ -90,7 +99,8 @@ function promote_operation(
     α::Type{<:Number},
 )
     return error(
-        "Operation `$op` between `$A` and `$α` is not allowed. You should use broadcast.",
+        "Operation `$op` between `$A` and `$α` is not allowed. You should " *
+        "use broadcast.",
     )
 end
 function promote_operation(
@@ -99,7 +109,8 @@ function promote_operation(
     A::Type{<:Array},
 )
     return error(
-        "Operation `$op` between `$α` and `$A` is not allowed. You should use broadcast.",
+        "Operation `$op` between `$α` and `$A` is not allowed. You should " *
+        "use broadcast.",
     )
 end
 
@@ -152,6 +163,7 @@ function operate end
 #     API without altering `x` and `y`. If it is not the case, implement a
 #     custom `operate` method.
 operate(::typeof(-), x) = -x
+
 function operate(
     op::Union{typeof(+),typeof(*),AddSubMul,typeof(add_dot)},
     x,
@@ -160,17 +172,21 @@ function operate(
 ) where {N}
     return op(x, y, args...)
 end
+
 operate(op::Union{typeof(-),typeof(/)}, x, y) where {N} = op(x, y)
+
[…]
