### 2022-02-07 772c4e0
Style updates for interface.jl (#146)
* Style updates for interface.jl

* Update interface.jl
