### 2022-02-07 1077a6e
WIP: investigate performance issues
