### 2022-02-07 e3268b4
reducing allocations always feels great ;)
