+++ test/broadcast.jl
@@ -28,7 +28,7 @@ end
     if VERSION >= v"1.5"
         # FIXME This should not allocate but I couldn't figure out where these
         #       240 come from.
-        alloc_test(() -> MA.broadcast!!(+, a, b), 240)
+        alloc_test(() -> MA.broadcast!!(+, a, b), 80)
         alloc_test(() -> MA.broadcast!!(+, a, c), 0)
     end
 end
