### 2022-02-07 b6ab947
Update rational.jl
