### 2022-02-08 3eda167
Improve performance of LinearAlgebra.dot (#147)
