### 2022-02-08 f5c54dd
Add Aqua.jl as a CI job
