+++ src/Test/config.jl
@@ -11,7 +11,7 @@ macro test_suite(setname, subsets = false)
             function $testname(
                 args...;
                 exclude::Vector{String} = String[],
-            ) where {T}
+            )
                 for (name, f) in $testdict
                     if name in exclude
                         continue
