+++ .github/workflows/aqua.yml
@@ -1,0 +1,23 @@
+name: aqua-lint
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    types: [opened, synchronize, reopened]
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: julia-actions/setup-julia@latest
+        with:
+          version: '1'
+      - uses: actions/checkout@v1
+      - name: Aqua
+        shell: julia --color=yes {0}
+        run: |
+          using Pkg
+          Pkg.add(PackageSpec(name="Aqua"))
+          Pkg.develop(PackageSpec(path=pwd()))
+          using MutableArithmetics, Aqua
+          Aqua.test_all(MutableArithmetics; ambiguities = false)
