### 2022-02-08 8f56f20
Add Aqua.jl as a CI job (#148)
