+++ src/Test/config.jl
@@ -7,20 +7,15 @@ macro test_suite(setname, subsets = false)
         runtest = :(f(args...))
     end
     return esc(
-        :(
-            function $testname(
-                args...;
-                exclude::Vector{String} = String[],
-            ) where {T}
-                for (name, f) in $testdict
-                    if name in exclude
-                        continue
-                    end
-                    @testset "$name" begin
-                        $runtest
-                    end
+        :(function $testname(args...; exclude::Vector{String} = String[])
+            for (name, f) in $testdict
+                if name in exclude
+                    continue
+                end
+                @testset "$name" begin
+                    $runtest
                 end
             end
-        ),
+        end),
     )
 end
