### 2022-02-08 d194f98
Fix for StaticArrays in multiplication (#142)
