+++ src/interface.jl
@@ -306,15 +306,14 @@ end
 Modify the value of `output` to be equal to the value of `op(args...)`. Can
 only be called if `mutability(output, op, args...)` returns [`IsMutable`](@ref).
 
-If `output === args[i]` for some `i`,
-* The user should expect to get an error. `operate!!` or `operate!` should be used instead.
-* Any method not supporting this case should throw an error.
+If `output === args[i]` for some `i`, this function may throw an error. Use
+`operate!!` or `operate!` instead.
 
-For instance, in DynamicPolynomials, `operate_to!(p, +, p, q)` throws an
-error because otherwise, the algorithm would fill `p` while iterating over the
-terms of `p` and `q` hence it will never terminate. On the other hand
-`operate!(+, p, q)` uses a different algorithm that efficiently inserts
-the terms of `q` in the sorted list of terms of `p` with minimal displacement.
+For example, in DynamicPolynomials, `operate_to!(p, +, p, q)` throws an error
+because otherwise, the algorithm would fill `p` while iterating over the terms
+of `p` and `q` hence it will never terminate. On the other hand
+`operate!(+, p, q)` uses a different algorithm that efficiently inserts the
+terms of `q` in the sorted list of terms of `p` with minimal displacement.
 """
 function operate_to!(output, op::F, args::Vararg{Any,N}) where {F<:Function,N}
     return operate_to_fallback!(
