### 2022-02-09 d696c6a
Prep for v1.0 release and drop support for pre Julia 1.6 versions
