+++ test/broadcast.jl
@@ -25,6 +25,6 @@ end
     @test y == 5
     # FIXME This should not allocate but I couldn't figure out where these
     #       240 come from.
-    alloc_test(() -> MA.broadcast!!(+, a, b), 240)
+    alloc_test(() -> MA.broadcast!!(+, a, b), 30 * sizeof(Int))
     alloc_test(() -> MA.broadcast!!(+, a, c), 0)
 end
