### 2022-02-15 bcafec7
Prep for v1.0 release and drop support for pre Julia 1.6 versions (#150)
