+++ test/broadcast.jl
@@ -10,10 +10,8 @@ const MA = MutableArithmetics
     @test a == [4, 5]
     # Need to have immutable structs that contain references to be allocated on
     # the stack: https://github.com/JuliaLang/julia/pull/33886
-    if VERSION >= v"1.5"
-        alloc_test(() -> MA.broadcast!!(+, a, b), 0)
-        alloc_test(() -> MA.broadcast!!(+, a, c), 0)
-    end
+    alloc_test(() -> MA.broadcast!!(+, a, b), 0)
+    alloc_test(() -> MA.broadcast!!(+, a, c), 0)
 end
 @testset "BigInt" begin
     x = BigInt(1)
@@ -25,10 +23,8 @@ end
     @test a == [4, 5]
     @test x == 4
     @test y == 5
-    if VERSION >= v"1.5"
-        # FIXME This should not allocate but I couldn't figure out where these
-        #       240 come from.
-        alloc_test(() -> MA.broadcast!!(+, a, b), 240)
-        alloc_test(() -> MA.broadcast!!(+, a, c), 0)
-    end
+    # FIXME This should not allocate but I couldn't figure out where these
+    #       240 come from.
+    alloc_test(() -> MA.broadcast!!(+, a, b), 30 * sizeof(Int))
+    alloc_test(() -> MA.broadcast!!(+, a, c), 0)
 end
