### 2022-05-12 3ea13af
add triangular case
