+++ src/Test/int.jl
@@ -1,3 +1,9 @@
+# Copyright (c) 2019 MutableArithmetics.jl contributors
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v.2.0. If a copy of the MPL was not distributed with this file, You can obtain
+# one at http://mozilla.org/MPL/2.0/.
+
 function int_add_test(::Type{T}) where {T}
     @testset "add_to! / add!" begin
         @test MA.mutability(T, +, T, T) isa MA.IsMutable
