### 2022-05-14 ebcaba5
add `mul!` for `AbstractTriangular` (#153)
