### 2022-05-15 3c71f83
Fix Vector-AdjVector product
