+++ src/implementations/LinearAlgebra.jl
@@ -199,6 +199,13 @@ function promote_array_mul(
 end
 
 function promote_array_mul(
+    ::Type{<:AbstractVector{S}},
+    ::Type{<:LinearAlgebra.Adjoint{T,<:AbstractVector{T}}},
+) where {S,T}
+    return Matrix{promote_sum_mul(S, T)}
+end
+
+function promote_array_mul(
     ::Type{<:AbstractMatrix{S}},
     ::Type{<:AbstractVector{T}},
 ) where {S,T}
