### 2022-05-16 788364d
Fix tests for PR#153 (#156)
