### 2022-06-08 a94ecc3
Add integration test for CI and fix JuMP
