+++ .github/workflows/downstream.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       matrix:
         julia-version: ['1.6', '1']
-        os: ubuntu-latest
+        os: ['ubuntu-latest']
         package:
           - {user: jump-dev, repo: JuMP.jl}
     steps:
