### 2022-06-08 4b69501
Add integration test for CI and fix JuMP (#160)
