+++ src/broadcast.jl
@@ -89,7 +89,7 @@ function broadcast_mutability(x, op, args::Vararg{Any,N}) where {N}
     return broadcast_mutability(typeof(x), op, typeof.(args)...)
 end
 
-_checked_size(s, x::AbstractArray) = size(x) == s
+_checked_size(s, x::AbstractArray) = length(x) == s
 _checked_size(::Any, ::Any) = true
 _checked_size(::Any, ::Tuple{}) = true
 function _checked_size(s, x::Tuple)
@@ -102,13 +102,15 @@ end
 # happen during broadcasting since we'll either need to return a different size
 # to `x`, or multiple copies of an argument will be used for different parts of
 # `x`. To simplify, let's just return `IsNotMutable` if the sizes are different,
-# which will be slower but correct.
+# which will be slower but correct. This is slightly complicated by the fact
+# that some AbstractArray do not support `size`, so we check with `length`
+# instead. If the `size`s are different, a later error will be thrown.
 function broadcast_mutability(
     x::AbstractArray,
     op,
     args::Vararg{Any,N},
 ) where {N}
-    if !_checked_size(size(x), args)::Bool
+    if !_checked_size(length(x), args)::Bool
         return IsNotMutable()
     end
     return broadcast_mutability(typeof(x), op, typeof.(args)...)
