+++ paper/paper.tex
@@ -6,6 +6,8 @@
 
 \newcommand{\ma}{MutableArithmetics}
 
+\addlitjlmacros{@rewrite}{@rewrite}{8}
+
 \begin{document}
 
 \input{header}
@@ -59,7 +61,7 @@ The section is organized into four subsections that describe the need of four ke
 \label{sec:may_mutate}
 Consider the task of summing the elements of a vector.
 By default, Julia \lstinline|sum| function will compute the sum with a code equivalent to the following:
-\begin{lstlisting}[language = Julia]
+\begin{jllisting}
 function sum(x::Vector)
     acc = zero(eltype(x))
     for el in x
@@ -67,7 +69,7 @@ function sum(x::Vector)
     end
     return acc
 end
-\end{lstlisting}
+\end{jllisting}
 If the type of the elements of \lstinline|x| is \lstinline|BigInt|, it is more efficient to replace the line
 \lstinline|acc = acc + el| by the line
 \lstinline|Base.GMP.MPZ.add!(acc, el)|.
@@ -81,7 +83,7 @@ the time complexity order is identical: $\Theta(nm)$ in both cases where $n$ is 
 We now consider a mutable element type for which exploiting mutability affects the time complexity.
 Consider a type \lstinline|SymbolicVariable| representing a symbolic variable and the following types representing a linear combinations of these variables with coefficients of type \lstinline|T|.
 This example encapsulates for instance JuMP affine expressions~\cite{dunning2017jump}, MOI affine functions~\cite{legat2021mathoptinterface}, polynomials (univariate~\cite{verzani2021polynomials} or multivariate~\cite{legat2021multivariatepolynomials}) or symbolic sums~\cite{gowda2021high}.
-\begin{lstlisting}[language = Julia]
+\begin{jllisting}
 struct Term{T}
     coef::T
     sym::SymbolicVariable
@@ -91,7 +93,7 @@ struct Sum{T}
 end
 Base.:+(s::Sum, t::Term) = Sum(push!(copy(s.terms), t))
 Base.zero(::Type{Term{T}}) where {T} = Sum(Term{T}[])
-\end{lstlisting}
+\end{jllisting}
 Calling \lstinline|sum| on a vector of $n$ \lstinline|Term{T}| has a time complexity $\Theta(n^2)$.
 Indeed, when calling \lstinline|acc + el| where \lstinline|acc| contains the sum of the first \lstinline|k| terms and \lstinline|el| is the $(k+1)$th term,
 the result cannot mutate \lstinline|acc.terms| and the copy of \lstinline|acc.terms| has time complexity $\Theta(k)$.
@@ -103,7 +105,7 @@ By default, \lstinline|add!!| would fallback to calling \lstinline|+|
 so that a method calling \lstinline|add!!| would both exploit the mutability
 of mutable types but would also work for non-mutable types.
 For our example, an implementation could be:
-\begin{lstlisting}[language = Julia]
+\begin{jllisting}
 function sum(x)
     acc = zero(eltype(x))
     for el in x
@@ -117,7 +119,7 @@ function add!!(s::Sum, t::Term)
     push!(s.terms, t)
     return s
 end
-\end{lstlisting}
+\end{jllisting}
 Note that the time complexity of the sum of $n$ \lstinline|Term| is now $\Theta(n)$.
 
 Julia implements a specialized method for computing the sum of \lstinline|BigInt|s that uses \lstinline|Base.GMP.MPZ.add!|.
@@ -134,12 +136,12 @@ When writing a code that should mutate the first argument of an operation,
[…]
