+++ src/interface.jl
@@ -105,7 +105,10 @@ function promote_operation(op::F, args::Vararg{Type,N}) where {F<:Function,N}
     return promote_operation_fallback(op, args...)
 end
 
-promote_operation(::F) where {F} = Any
+# Fix ambiguities identified by Aqua.jl
+promote_operation(::typeof(-)) = Any
+promote_operation(::typeof(+)) = Any
+promote_operation(::typeof(*)) = Any
 
 # Helpful error for common mistake
 function promote_operation(
@@ -369,7 +372,9 @@ function operate!(op::F, args::Vararg{Any,N}) where {F<:Function,N}
 end
 
 buffer_for(::F, args::Vararg{Type,N}) where {F<:Function,N} = nothing
-buffer_for(::F) where {F<:Function} = nothing
+
+# Fix ambiguity identified by Aqua.jl
+buffer_for(::AddSubMul) = nothing
 
 function buffered_operate_to_fallback!(
     ::IsNotMutable,
