+++ .github/workflows/aqua.yml
@@ -21,4 +21,17 @@ jobs:
           Pkg.develop(PackageSpec(path=pwd()))
           using MutableArithmetics, Aqua
           # Skip ambiguity test for now
-          Aqua.test_all(MutableArithmetics; ambiguities = false)
+          Aqua.test_all(
+            MutableArithmetics;
+            ambiguities = (
+              exclude = [
+                # TODO(odow): These ones are suspect, and may be bugs.
+                MutableArithmetics.LinearAlgebra.mul!,
+                # These ones are probably safe to ignore because there are
+                # only called internally.
+                MutableArithmetics.buffered_operate_fallback!,
+                MutableArithmetics.buffered_operate_to_fallback!,
+                MutableArithmetics.promote_operation_fallback,
+              ],
+            ),
+          )
