### 2022-08-12 35948ab
Fix promote_operation for irrational numbers
