### 2022-08-12 d86ca06
Update `promote_operation` for `Irrational`
