### 2022-08-12 3488eae
Add test for `promote_operation` with user-defined `Irrational` input
