### 2022-08-13 c54455a
Make `_instantiate(::Type{Irrational{S}})` return corresponding value
