+++ src/interface.jl
@@ -28,9 +28,8 @@ end
 _instantiate_zero(::Type{S}) where {S} = zero(S)
 _instantiate_oneunit(::Type{S}) where {S} = oneunit(S)
 
-# a helper function for promotion
-# the specific return value doesn't matter
-_instantiate(::Type{T}) where {T<:Irrational} = π
+# this is valid because Irrational numbers are defined in global scope as const
+_instantiate(::Type{S}) where {S<:Irrational} = S()
 _instantiate_zero(::Type{S}) where {S<:AbstractIrrational} = _instantiate(S)
 _instantiate_oneunit(::Type{S}) where {S<:AbstractIrrational} = _instantiate(S)
 
