### 2022-08-19 ec17103
Support user-defined `Irrational` in `promote_operation` (#166)
* Update `promote_operation` for `Irrational`

* Add test for `promote_operation` with user-defined `Irrational` input

* Format

* Make `_instantiate(::Type{Irrational{S}})` return corresponding value
