+++ src/interface.jl
@@ -27,13 +27,9 @@ end
 
 _instantiate_zero(::Type{S}) where {S} = zero(S)
 _instantiate_oneunit(::Type{S}) where {S} = oneunit(S)
-# This would allocate in addition to not working for constants defined outside Base:
-# `_instantiate_(::Type{Irrational{S}}) where {S} = eval(Expr(:., Base.MathConstants, QuoteNode(S)))`
-# so we do
-_instantiate(::Type{Irrational{:π}}) = Base.MathConstants.π
-_instantiate(::Type{Irrational{:φ}}) = Base.MathConstants.φ
-_instantiate(::Type{Irrational{:γ}}) = Base.MathConstants.γ
-_instantiate(::Type{Irrational{:ℯ}}) = Base.MathConstants.ℯ
+
+# this is valid because Irrational numbers are defined in global scope as const
+_instantiate(::Type{S}) where {S<:Irrational} = S()
 _instantiate_zero(::Type{S}) where {S<:AbstractIrrational} = _instantiate(S)
 _instantiate_oneunit(::Type{S}) where {S<:AbstractIrrational} = _instantiate(S)
 
