### 2022-09-21 731f211
Fix promote_operation for irrational numbers (#165)
* Fix promote_operation for irrational numbers

* Fixes

* Support user-defined `Irrational` in `promote_operation` (#166)

* Update `promote_operation` for `Irrational`

* Add test for `promote_operation` with user-defined `Irrational` input

* Format

* Make `_instantiate(::Type{Irrational{S}})` return corresponding value

Co-authored-by: Bowen S. Zhu <bszhu@fas.harvard.edu>
