### 2022-10-06 54f0bde
Simplify rationals on addition and substraction
