### 2022-10-07 80b95af
Simplify rationals on addition and substraction (#168)
* Simplify rationals on addition and substraction

* Fix

* Fix

* Fix format

* aqua-lint fix
