+++ src/new_rewrite.jl
@@ -1,0 +1,253 @@
+# Copyright (c) 2019 MutableArithmetics.jl contributors
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v.2.0. If a copy of the MPL was not distributed with this file, You can obtain
+# one at http://mozilla.org/MPL/2.0/.
+
+module MutableArithmetics2
+
+import ..MutableArithmetics
+
+const MA = MutableArithmetics
+
+"""
+    @rewrite(expr)
+
+Rewrites the expression `expr` to use mutable arithmetics.
+
+For a non-macro version, see [`rewrite_and_return`](@ref).
+"""
+macro rewrite(expr)
+    return rewrite_and_return(expr)
+end
+
+"""
+    rewrite_and_return(expr) -> Expr
+
+Rewrites the expression `expr` to use mutable arithmetics.
+"""
+function rewrite_and_return(expr)
+    stack = quote end
+    root, _ = _rewrite(stack, expr)
+    return quote
+        let
+            $stack
+            $root
+        end
+    end
+end
+
+"""
+    rewrite(expr) -> Tuple{Symbol,Expr}
+
+Rewrites the expression `expr` to use mutable arithmetics. Returns
+`(variable, code)` comprised of a `gensym`'d variable equivalent to `expr` and
+the code necessary to create the variable.
+"""
+function rewrite(expr)
+    variable = gensym()
+    code = rewrite_and_return(expr)
+    return variable, :($variable = $code)
+end
+
+"""
+    _rewrite(stack::Expr, expr::T)::Tuple{Any,Bool}
+
+This method is the heart of the rewrite logic. It converts `expr` into a mutable
+equivalent, places any intermediate calculations onto `stack`, and returns a
+tuple containing the return value---which is either `expr` or a `gensym`ed
+variable equivalent to `expr`---and a boolean flag that indicates whether the
+return value can be mutated by future callers.
+"""
+function _rewrite end
+"""
+    _rewrite(::Expr, x)
+
+A generic fallback. Given a type `x` we return it without mutation. In addition,
+this type should not be mutated by future callers.
+"""
+_rewrite(::Expr, x) = esc(x), false
+
+"""
+    _rewrite(::Expr, x::Number)
+
+If `x` is a `Number` at macro expansion time, it _must_ be a constant literal.
+We return `x` without mutation, but we return `true` because other callers may
+mutate the value without fear. Put aother way, they don't need to wrap the value
+in `copy_if_mutable(x)` before using it as the first argument to `operate!!`.
+
+This most commonly happens in situations like `x^2`.
+"""
+_rewrite(::Expr, x::Number) = x, true
+
+"""
+    _rewrite(stack::Expr, expr::Expr)
+
+This method is the heart of the rewrite logic. It converts `expr` into a mutable
+equivalent.
+"""
+function _rewrite(stack::Expr, expr::Expr)
+    if !Meta.isexpr(expr, :call)
+        # In situations like `x[i]`, we do not attempt to rewrite. Return `expr`
+        # and don't let future callers mutate.
+        return esc(expr), false
+    elseif Meta.isexpr(expr, :call, 1)
+        # A zero-argument function
+        return esc(expr), false
+    elseif Meta.isexpr(expr, :call, 2) && (
+        Meta.isexpr(expr.args[2], :generator) ||
+        Meta.isexpr(expr.args[2], :flatten)
+    )
+        # This is a generator expression like `sum(i for i in args)`. Generators
+        # come in two forms: `sum(i for i=I, j=J)` or `sum(i for i=I for j=J)`.
+        # The latter is a `:flatten` expression and needs additional handling,
+        # but we delay this complexity for _rewrite_generator.
+        if expr.args[1] in (:sum, :Σ, :∑)
[…]
