### 2022-11-07 3deb169
WIP: rewrite experiments
Fix rewrite with views

Update

Support broadcasting

Fix handling of *

Fix generators

Add tests for repeated sums

Fix formatting

Updates

Place new rewrite behind opt-in kwarg

More coverage

Update docstrings
