### 2022-11-07 22ddc6e
Fix implementation of fused_map_reduce
