+++ test/dispatch.jl
@@ -9,18 +9,16 @@ using LinearAlgebra
 # Tests that the calls are correctly redirected to the mutable calls
 # by checking allocations
 function dispatch_tests(::Type{T}) where {T}
-    buffer = zero(T)
-    a = one(T)
-    b = one(T)
-    c = one(T)
     x = convert.(T, [1, 2, 3])
-    # Need to allocate 1 BigInt for the result and one for the buffer
-    alloc_test(() -> MA.fused_map_reduce(MA.add_mul, x, x), 2BIGINT_ALLOC)
-    alloc_test(() -> MA.fused_map_reduce(MA.add_dot, x, x), 2BIGINT_ALLOC)
+    # Need to allocate 1 BigInt for the result and one for the buffer, plus two
+    # zero(T) for determining the eltype of the result.
+    cost = Sys.WORD_SIZE == 64 ? (2 * 48 + 2 * 40) : (2 * 24 + 2 * 20)
+    alloc_test(() -> MA.fused_map_reduce(MA.add_mul, x, x), cost)
+    alloc_test(() -> MA.fused_map_reduce(MA.add_dot, x, x), cost)
     if T <: MA.AbstractMutable
-        alloc_test(() -> x'x, 2BIGINT_ALLOC)
-        alloc_test(() -> transpose(x) * x, 2BIGINT_ALLOC)
-        alloc_test(() -> LinearAlgebra.dot(x, x), 2BIGINT_ALLOC)
+        alloc_test(() -> x'x, cost)
+        alloc_test(() -> transpose(x) * x, cost)
+        alloc_test(() -> LinearAlgebra.dot(x, x), cost)
     end
 end
 
@@ -29,4 +27,17 @@ end
     # On `DummyBigInt` allocates more on previous releases of Julia
     # as it's dynamically allocated
     dispatch_tests(DummyBigInt)
+
+    @testset "dot non-concrete vector" begin
+        x = [5.0, 6.0]
+        y = Vector{Union{Float64,String}}(x)
+        @test MA.operate(LinearAlgebra.dot, x, y) == LinearAlgebra.dot(x, y)
+        @test MA.operate(*, x', y) == x' * y
+    end
+
+    @testset "dot vector of vectors" begin
+        x = [5.0, 6.0]
+        z = [x, x]
+        @test MA.operate(LinearAlgebra.dot, z, z) == LinearAlgebra.dot(z, z)
+    end
 end
