### 2022-11-07 aa6d20a
Fix JuMP tests
