+++ src/reduce.jl
@@ -26,12 +26,27 @@ map_op(::AddSubMul) = *
 
 map_op(::typeof(add_dot)) = LinearAlgebra.dot
 
+# We need a generated function here to help with type stability.
+@generated function _accumulator(f::F, op::O, args::Vararg{Any,N}) where {F,O,N}
+    expr = Expr(:call, :(map_op(op)))
+    for i in 1:N
+        push!(expr.args, :(zero(f(args[$i]))))
+    end
+    return expr
+end
+
 function fused_map_reduce(op::F, args::Vararg{Any,N}) where {F<:Function,N}
     _check_same_length(args...)
     if length(args[1]) == 0
-        return map_op(op)(zero.(eltype.(args))...)
+        # If there are no arguments, we need to use the eltype of each vector to
+        # infer the result type.
+        return _accumulator(eltype, op, args...)
     end
-    accumulator = map_op(op)(zero.(first.(args))...)
+    # If there are arguments, we use the first element. We don't use the eltype
+    # because non-concrete vectors might have an eltype that isn't amenable to
+    # zero(T). Ideally, this would be an error. But LinearAlgebra.dot supports
+    # it so we do too.
+    accumulator = _accumulator(first, op, args...)
     T = typeof(accumulator)
     buffer = buffer_for(op, T, eltype.(args)...)
     for I in zip(eachindex.(args)...)
