### 2022-11-08 8f526b3
Fix rewrite with views
