### 2022-11-08 01d8066
WIP: experiment replacing rewrite with MutableArithmetics2
