+++ src/rewrite.jl
@@ -102,171 +102,6 @@ Base.length(::Zero) = 1
 Base.iterate(z::Zero) = (z, nothing)
 Base.iterate(::Zero, ::Nothing) = nothing
 
-# See `JuMP._try_parse_idx_set`
-function _try_parse_idx_set(arg::Expr)
-    # [i=1] and x[i=1] parse as Expr(:vect, Expr(:(=), :i, 1)) and
-    # Expr(:ref, :x, Expr(:kw, :i, 1)) respectively.
-    if arg.head === :kw || arg.head === :(=)
-        @assert length(arg.args) == 2
-        return true, arg.args[1], arg.args[2]
-    elseif isexpr(arg, :call) && arg.args[1] === :in
-        # FIXME It seems not to be called anymore on Julia v1.0 as `i in idx`
-        #       is parsed as `i = idx`.
-        return true, arg.args[2], arg.args[3]
-    else
-        # FIXME When is this called ?
-        return false, nothing, nothing
-    end
-end
-
-function _parse_idx_set(arg::Expr)
-    parse_done, idxvar, idxset = _try_parse_idx_set(arg)
-    if !parse_done
-        error("Invalid syntax: $arg")
-    end
-    return idxvar, idxset
-end
-
-"""
-    rewrite_generator(expr::Expr, inner::Function)
-
-Rewrites the generator statements `expr` and returns a properly nested for loop
-with nested filters as specified.
-
-# Examples
-
-```jldoctest
-julia> using MutableArithmetics
-
-julia> MutableArithmetics.rewrite_generator(:(i for i in 1:2 if isodd(i)), i -> :(\$i + 1))
-:(for \$(Expr(:escape, :(i = 1:2)))
-      if \$(Expr(:escape, :(isodd(i))))
-          i + 1
-      end
-  end)
-```
-"""
-function rewrite_generator(ex, inner)
-    # `i + j for i in 1:2 for j in 1:2` is a `flatten` expression
-    if isexpr(ex, :flatten)
-        return rewrite_generator(ex.args[1], inner)
-    elseif !isexpr(ex, :generator)
-        return inner(ex)
-    end
-    # `i + j for i in 1:2, j in 1:2` is a `generator` expression
-    function itrsets(sets)
-        if isa(sets, Expr)
-            return sets
-        elseif length(sets) == 1
-            return sets[1]
-        else
-            return Expr(:block, sets...)
-        end
-    end
-
-    idxvars = []
-    if isexpr(ex.args[2], :filter) # if condition
-        loop = Expr(
-            :for,
-            esc(itrsets(ex.args[2].args[2:end])),
-            Expr(
-                :if,
-                esc(ex.args[2].args[1]),
-                rewrite_generator(ex.args[1], inner),
-            ),
-        )
-        for idxset in ex.args[2].args[2:end]
-            idxvar, s = _parse_idx_set(idxset)
-            push!(idxvars, idxvar)
-        end
-    else
-        loop = Expr(
-            :for,
-            esc(itrsets(ex.args[2:end])),
-            rewrite_generator(ex.args[1], inner),
-        )
-        for idxset in ex.args[2:end]
-            idxvar, s = _parse_idx_set(idxset)
-            push!(idxvars, idxvar)
-        end
-    end
-    return loop
-end
-
-# See `JuMP._is_sum`
-_is_sum(s::Symbol) = (s == :sum) || (s == :∑) || (s == :Σ)
-
-function _parse_generator(
-    vectorized::Bool,
-    minus::Bool,
-    inner_factor::Expr,
-    current_sum::Union{Nothing,Symbol},
-    left_factors,
-    right_factors,
-    new_var = gensym(),
-)
-    @assert isexpr(inner_factor, :call)
[…]
