### 2022-11-08 70447e5
Re-enable rewrite_gennerator
