### 2022-11-14 c2bde72
Use _concrete_eltype
