### 2022-11-15 201aafb
Fix implementation of fused_map_reduce (#171)
