+++ test/dispatch.jl
@@ -29,4 +29,17 @@ end
     # On `DummyBigInt` allocates more on previous releases of Julia
     # as it's dynamically allocated
     dispatch_tests(DummyBigInt)
+
+    @testset "dot non-concrete vector" begin
+        x = [5.0, 6.0]
+        y = Vector{Union{Float64,String}}(x)
+        @test MA.operate(LinearAlgebra.dot, x, y) == LinearAlgebra.dot(x, y)
+        @test MA.operate(*, x', y) == x' * y
+    end
+
+    @testset "dot vector of vectors" begin
+        x = [5.0, 6.0]
+        z = [x, x]
+        @test MA.operate(LinearAlgebra.dot, z, z) == LinearAlgebra.dot(z, z)
+    end
 end
