+++ src/reduce.jl
@@ -36,9 +36,11 @@ function promote_map_reduce(op::Function, args::Vararg{Any,N}) where {N}
     )
 end
 
+_concrete_eltype(x) = isempty(x) ? eltype(x) : typeof(first(x))
+
 function fused_map_reduce(op::F, args::Vararg{Any,N}) where {F<:Function,N}
     _check_same_length(args...)
-    T = promote_map_reduce(op, eltype.(args)...)
+    T = promote_map_reduce(op, _concrete_eltype.(args)...)
     accumulator = neutral_element(reduce_op(op), T)
     buffer = buffer_for(op, T, eltype.(args)...)
     for I in zip(eachindex.(args)...)
