### 2022-11-15 4a05f70
move_factors_into_sums
