### 2022-11-15 1b94a2a
Add tests for other forms of sum
