### 2022-11-15 ed097d1
Fix formatting and add tests
