### 2022-11-16 fefb084
Fix rewrite of * when not mutable
