### 2022-11-17 da4a81c
Add tests and fix for SparseArray matrix multiplication
