### 2022-11-17 f8f8c4f
Fix promotion fallbacks with non-concrete types
