### 2022-11-17 b46f175
Add tests and fix for SparseArray matrix multiplication (#173)
