+++ Project.toml
@@ -14,6 +14,7 @@ julia = "1.6"
 
 [extras]
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
 [targets]
-test = ["OffsetArrays"]
+test = ["OffsetArrays", "Random"]
