### 2022-11-17 02a2ca5
Fix methods and add tests for SparseArrays.jl
