### 2022-11-18 aef52ab
dispatch.jl: simplify number * sparse matrix
