### 2022-11-18 eb3972a
dispatch.jl: remove work-arounds for Julia 1.0
