### 2022-11-20 352aa8f
dispatch.jl: remove work-arounds for Julia 1.0 (#180)
