### 2022-11-20 4895637
Switch to Base.isconcretetype
