### 2022-11-20 c589ecb
dispatch.jl: simplify overloads for SparseMatrix * AbstractArray
