+++ test/matmul.jl
@@ -64,29 +64,6 @@ end
     @test MA.operate(convert, Int, 1) === 1
 end
 
-const EXPECTED_ERROR = string(
-    "Cannot multiply a `Matrix{NoProdMutable}` with a ",
-    "`Matrix{NoProdMutable}` because the sum of the product of a ",
-    "`NoProdMutable` and a `NoProdMutable` could not be inferred so a ",
-    "`Matrix{Union{}}` allocated to store the output of the ",
-    "multiplication instead of a `Matrix{$(Int)}`.",
-)
-
-struct NoProdMutable <: MA.AbstractMutable end
-function MA.promote_operation(
-    ::typeof(*),
-    ::Type{NoProdMutable},
-    ::Type{NoProdMutable},
-)
-    return Int # Dummy result just to test error message
-end
-
-function unsupported_product()
-    A = [NoProdMutable() for i in 1:2, j in 1:2]
-    err = ErrorException(EXPECTED_ERROR)
-    @test_throws err A * A
-end
-
 @testset "Errors" begin
     @testset "`promote_op` error" begin
         AT = CustomArray{Int,3}
