+++ src/dispatch.jl
@@ -92,251 +92,6 @@ function LinearAlgebra.diagm(x::AbstractVector{<:AbstractMutable})
     return LinearAlgebra.diagm(0 => copyto!(similar(x, ZeroType), x))
 end
 
-################################################################################
-# Interception of Base's matrix/vector arithmetic machinery
-#
-# Redirect calls with `eltype(ret) <: AbstractMutable` to `_mul!` to replace it
-# with an implementation more efficient than `generic_matmatmul!` and
-# `generic_matvecmul!` since it takes into account the mutability of the
-# arithmetic. We need `args...` because SparseArrays` also gives `α` and `β`
-# arguments.
-
-function _mul!(output, A, B, α, β)
-    # See SparseArrays/src/linalg.jl
-    if !isone(β)
-        if iszero(β)
-            operate!(zero, output)
-        else
-            rmul!(output, scaling(β))
-        end
-    end
-    return operate!(add_mul, output, A, B, scaling(α))
-end
-
-function _mul!(output, A, B, α)
-    operate!(zero, output)
-    return operate!(add_mul, output, A, B, scaling(α))
-end
-
-# LinearAlgebra uses `Base.promote_op(LinearAlgebra.matprod, ...)` to try to
-# infere the return type. If the operation is not supported, it returns
-# `Union{}`.
-function _mul!(output::AbstractArray{Union{}}, A, B)
-    # Normally, if the product is not supported, this should redirect to
-    # `MA.promote_operation(*, ...)` which redirects to
-    # `zero(...) * zero(...)` which should throw an appropriate error.
-    # For example, in JuMP, it would say that you cannot multiply quadratic
-    # expressions with an affine expression for instance.
-    ProdType = promote_array_mul(typeof(A), typeof(B))
-    # If we arrived here, it means that we have found a type for `output`, even
-    # if LinearAlgebra couldn't. This is most probably a but so let's provide
-    # extensive information to help debugging.
-    return error(
-        "Cannot multiply a `$(typeof(A))` with a `$(typeof(B))` because the " *
-        "sum of the product of a `$(eltype(A))` and a `$(eltype(B))` could " *
-        "not be inferred so a `$(typeof(output))` allocated to store the " *
-        "output of the multiplication instead of a `$ProdType`.",
-    )
-end
-
-_mul!(output, A, B) = operate_to!(output, *, A, B)
-
-function LinearAlgebra.mul!(
-    ret::AbstractMatrix{<:AbstractMutable},
-    A::AbstractVecOrMat,
-    B::AbstractVecOrMat,
-)
-    return _mul!(ret, A, B)
-end
-
-function LinearAlgebra.mul!(
-    ret::AbstractVector{<:AbstractMutable},
-    A::AbstractVecOrMat,
-    B::AbstractVector,
-)
-    return _mul!(ret, A, B)
-end
-
-function LinearAlgebra.mul!(
-    ret::AbstractVector{<:AbstractMutable},
-    A::LinearAlgebra.Transpose{<:Any,<:AbstractVecOrMat},
-    B::AbstractVector,
-)
-    return _mul!(ret, A, B)
-end
-
[…]
