+++ src/implementations/LinearAlgebra.jl
@@ -385,6 +385,12 @@ function operate(
     A::AbstractMatrix{S},
     B::AbstractVector{T},
 ) where {T,S}
+    # Only use the efficient in-place operate_to! if both arrays are
+    # concrete. Bad things can happen if S or T is abstract and we pick the
+    # wrong type for C.
+    if !(isconcretetype(S) && isconcretetype(T))
+        return A * B
+    end
     C = undef_array(promote_array_mul(typeof(A), typeof(B)), axes(A, 1))
     return operate_to!(C, *, A, B)
 end
@@ -394,6 +400,12 @@ function operate(
     A::AbstractMatrix{S},
     B::AbstractMatrix{T},
 ) where {T,S}
+    # Only use the efficient in-place operate_to! if both arrays are
+    # concrete. Bad things can happen if S or T is abstract and we pick the
+    # wrong type for C.
+    if !(isconcretetype(S) && isconcretetype(T))
+        return A * B
+    end
     C = undef_array(
         promote_array_mul(typeof(A), typeof(B)),
         axes(A, 1),
