### 2022-11-21 c715a0c
Add new rewrite that does not make strong assumptions on result type (#170)
