+++ src/rewrite_generic.jl
@@ -1,0 +1,280 @@
+# Copyright (c) 2019 MutableArithmetics.jl contributors
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v.2.0. If a copy of the MPL was not distributed with this file, You can obtain
+# one at http://mozilla.org/MPL/2.0/.
+
+# We need these two methods because we're changing how * is re-written.
+operate!(::typeof(*), x::AbstractArray{T}, y::T) where {T} = (x .*= y)
+operate!(::typeof(*), x::AbstractArray, y) = (x .= operate(*, x, y))
+
+"""
+    _rewrite_generic(stack::Expr, expr::T)::Tuple{Any,Bool}
+
+This method is the heart of the rewrite logic. It converts `expr` into a mutable
+equivalent, places any intermediate calculations onto `stack`, and returns a
+tuple containing the return value---which is either `expr` or a `gensym`ed
+variable equivalent to `expr`---and a boolean flag that indicates whether the
+return value can be mutated by future callers.
+"""
+function _rewrite_generic end
+
+"""
+    _rewrite_generic(::Expr, x)
+
+A generic fallback. Given a type `x` we return it without mutation. In addition,
+this type should not be mutated by future callers.
+"""
+_rewrite_generic(::Expr, x) = esc(x), false
+
+"""
+    _rewrite_generic(::Expr, x::Number)
+
+If `x` is a `Number` at macro expansion time, it _must_ be a constant literal.
+We return `x` without mutation, but we return `true` because other callers may
+mutate the value without fear. Put aother way, they don't need to wrap the value
+in `copy_if_mutable(x)` before using it as the first argument to `operate!!`.
+
+This most commonly happens in situations like `x^2`.
+"""
+_rewrite_generic(::Expr, x::Number) = x, true
+
+function _is_generator(expr)
+    return Meta.isexpr(expr, :call, 2) && Meta.isexpr(expr.args[2], :generator)
+end
+
+function _is_flatten(expr)
+    return Meta.isexpr(expr, :call, 2) && Meta.isexpr(expr.args[2], :flatten)
+end
+
+function _is_parameters(expr)
+    return Meta.isexpr(expr, :call, 3) && Meta.isexpr(expr.args[2], :parameters)
+end
+
+"""
+    _rewrite_generic(stack::Expr, expr::Expr)
+
+This method is the heart of the rewrite logic. It converts `expr` into a mutable
+equivalent.
+"""
+function _rewrite_generic(stack::Expr, expr::Expr)
+    if !Meta.isexpr(expr, :call)
+        # In situations like `x[i]`, we do not attempt to rewrite. Return `expr`
+        # and don't let future callers mutate.
+        return esc(expr), false
+    elseif Meta.isexpr(expr, :call, 1)
+        # A zero-argument function
+        return esc(expr), false
+    elseif _is_generator(expr) || _is_flatten(expr) || _is_parameters(expr)
+        if !(expr.args[1] in (:sum, :Σ, :∑))
+            # We don't know what this is. Return the expression and don't let
+            # future callers mutate.
+            return esc(expr), false
+        end
+        # This is a generator expression like `sum(i for i in args)`. Generators
+        # come in two forms: `sum(i for i=I, j=J)` or `sum(i for i=I for j=J)`.
+        # The latter is a `:flatten` expression and needs additional handling,
+        # but we delay this complexity for _rewrite_generic_generator.
+        if Meta.isexpr(expr.args[2], :parameters, 1) &&
+           Meta.isexpr(expr.args[2].args[1], :kw, 2) &&
+           expr.args[2].args[1].args[1] == :init
+            # sum(iter ; init) form!
+            root = gensym()
[…]
