### 2022-11-22 461955e
Support splatting in rewrite_generic.jl
