### 2022-11-22 dfebb83
Support splatting in rewrite_generic.jl (#185)
