+++ versions.js
@@ -1,10 +1,11 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.1",
   "v1.0",
   "v0.3",
   "v0.2",
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.0.5";
+var DOCUMENTER_NEWEST = "v1.1.0";
 var DOCUMENTER_STABLE = "stable";
