### 2023-01-03 2264a78
0-allocation add_mul for Rational{BigInt}
