+++ test/big.jl
@@ -48,8 +48,8 @@ end
         allocation_test(+, T, MA.add!!, MA.add_to!!, T <: Rational ? 168 : 0)
         allocation_test(-, T, MA.sub!!, MA.sub_to!!, T <: Rational ? 168 : 0)
         allocation_test(*, T, MA.mul!!, MA.mul_to!!, T <: Rational ? 240 : 0)
-        add_sub_mul_test(add_mul, T)
-        add_sub_mul_test(sub_mul, T)
+        add_sub_mul_test(MA.add_mul, T)
+        add_sub_mul_test(MA.sub_mul, T)
         if T <: Rational # https://github.com/jump-dev/MutableArithmetics.jl/issues/167
             allocation_test(
                 +,
