### 2023-01-05 3cf0d8b
buffered_operate_to! for BigFloat dot product
Implements Kahan-Babuška-Neumaier compensated summation for the dot
product.

Fixes #189
