### 2023-01-05 a75aaea
add BigFloat negation and absolute value
Adapted from -(x::BigFloat) in Base.MPFR in mpfr.jl

Fixes #190
Fixes #191
