+++ src/implementations/BigFloat.jl
@@ -78,6 +78,30 @@ function operate_to!(output::BigFloat, ::typeof(-), a::BigFloat, b::BigFloat)
     return output
 end
 
+function operate!(::typeof(-), x::BigFloat)
+    ccall(
+        (:mpfr_neg, :libmpfr),
+        Int32,
+        (Ref{BigFloat}, Ref{BigFloat}, Base.MPFR.MPFRRoundingMode),
+        x,
+        x,
+        Base.MPFR.ROUNDING_MODE[])
+    return x
+end
+
+# Base.abs
+
+function operate!(::typeof(Base.abs), x::BigFloat)
+    ccall(
+        (:mpfr_abs, :libmpfr),
+        Int32,
+        (Ref{BigFloat}, Ref{BigFloat}, Base.MPFR.MPFRRoundingMode),
+        x,
+        x,
+        Base.MPFR.ROUNDING_MODE[])
+    return x
+end
+
 # *
 
 promote_operation(::typeof(*), ::Type{BigFloat}, ::Type{BigFloat}) = BigFloat
