+++ test/bigfloat_neg_abs.jl
@@ -4,13 +4,17 @@
 # v.2.0. If a copy of the MPL was not distributed with this file, You can obtain
 # one at http://mozilla.org/MPL/2.0/.
 
-function neg_abs_test(f::F, x::BigFloat) where {F <: Function}
+function neg_abs_test(f::F, x::BigFloat) where {F<:Function}
     backup = x + 0
     output = MA.operate!(f, x)
     @test f(backup) == output == x
+    return
 end
 
-@testset "$input" for input in map(BigFloat, (-1.3, -1.0, -0.7, -0.0, 0.0, 0.3, 1.0, 1.7))
+@testset "$input" for input in map(
+    BigFloat,
+    (-1.3, -1.0, -0.7, -0.0, 0.0, 0.3, 1.0, 1.7),
+)
     @testset "-" begin
         neg_abs_test(-, input)
     end
