### 2023-01-06 513f87f
add BigFloat negation and absolute value (#192)
* add BigFloat negation and absolute value

Adapted from -(x::BigFloat) in Base.MPFR in mpfr.jl

Fixes #190
Fixes #191

* Apply suggestions from code review

Co-authored-by: Oscar Dowson <odow@users.noreply.github.com>
