### 2023-01-10 a2730c2
implement BigFloat negation and abs without ccall
