+++ src/implementations/SparseArrays.jl
@@ -262,9 +262,13 @@ function operate!(
     return ret
 end
 
+# Taken from the internal function `Base.top_set_bit`, added in
+# https://github.com/JuliaLang/julia/pull/47523
+_top_set_bit(x::Integer) = 8sizeof(x) - leading_zeros(x)
+
 # Taken from `SparseArrays.prefer_sort` added in Julia v1.1.
 function _prefer_sort(nz::Integer, m::Integer)
-    return m > 6 && 3 * SparseArrays.ilog2(nz) * nz < m
+    return m > 6 && 3 * _top_set_bit(nz) * nz < m
 end
 
 function operate!(
