### 2023-01-31 0e9dffc
SparseArrays: fix for nightly Julia, ilog2 -> top_set_bit (#195)
Issue: MutableArithmetics for SparseArrays is broken on nightly Julia
because SparseArrays.ilog2 doesn't exist anymore.

This adds the function _top_set_bit and uses it instead of ilog2.

See
https://github.com/JuliaSparse/SparseArrays.jl/pull/324
and
https://github.com/JuliaLang/julia/pull/47523
