### 2023-02-14 594afd9
0-allocation add_mul for Rational{BigInt} (#188)
* 0-allocation add_mul for Rational{BigInt}

* Fixes

* Fixes

* Fix
