+++ test/big.jl
@@ -23,18 +23,33 @@ function allocation_test(
     g = op(a, b)
     @test c === short_to(c, a, b)
     @test g == c
-    @test a === short(a, b)
-    @test g == a
-    alloc_test_le(() -> short(a, b), n)
-    return alloc_test_le(() -> short_to(c, a, b), n)
+    A = MA.copy_if_mutable(a)
+    @test A === short(A, b)
+    @test g == A
+    alloc_test_le(() -> short(A, b), n)
+    alloc_test_le(() -> short_to(c, a, b), n)
+    @test g == MA.buffered_operate!(nothing, op, MA.copy_if_mutable(a), b)
+    @test g == MA.buffered_operate_to!(nothing, c, op, a, b)
+    buffer = MA.buffer_for(op, typeof(a), typeof(b))
+    @test g == MA.buffered_operate_to!(buffer, c, op, a, b)
+    alloc_test(() -> MA.buffered_operate_to!(buffer, c, op, a, b), 0)
+    return
 end
 
+function add_sub_mul_test(op, T; a = T(2), b = T(3), c = T(4))
+    g = op(a, b, c)
+    @test g == MA.buffered_operate!(nothing, op, MA.copy_if_mutable(a), b, c)
+    buffer = MA.buffer_for(op, typeof(a), typeof(b), typeof(c))
+    return alloc_test(() -> MA.buffered_operate!(buffer, op, a, b, c), 0)
+end
 @testset "$T" for T in [BigInt, BigFloat, Rational{BigInt}]
     MA.Test.int_test(T)
     @testset "Allocation" begin
         allocation_test(+, T, MA.add!!, MA.add_to!!, T <: Rational ? 168 : 0)
         allocation_test(-, T, MA.sub!!, MA.sub_to!!, T <: Rational ? 168 : 0)
         allocation_test(*, T, MA.mul!!, MA.mul_to!!, T <: Rational ? 240 : 0)
+        add_sub_mul_test(MA.add_mul, T)
+        add_sub_mul_test(MA.sub_mul, T)
         if T <: Rational # https://github.com/jump-dev/MutableArithmetics.jl/issues/167
             allocation_test(
                 +,
