+++ src/implementations/Rational.jl
@@ -33,22 +33,78 @@ function operate!(::typeof(one), x::Rational)
     return x
 end
 
-# +
+function _buffered_divgcd(buffer, x, x_output, y, y_output)
+    operate_to!(buffer, gcd, x, y)
+    operate_to!(x_output, div, x, buffer)
+    operate_to!(y_output, div, y, buffer)
+    return
+end
+
+function _buffered_divgcd(buffer, x, y)
+    operate_to!(buffer, gcd, x, y)
+    if !isone(buffer)
+        operate!(div, x, buffer)
+        operate!(div, y, buffer)
+    end
+    return
+end
+
+_buffered_simplify(buffer, x::Rational) = _buffered_divgcd(buffer, x.num, x.den)
+
+# + and -
 
 function promote_operation(
-    ::typeof(+),
+    ::Union{typeof(+),typeof(-)},
     ::Type{Rational{S}},
     ::Type{Rational{T}},
 ) where {S,T}
     return Rational{promote_sum_mul(S, T)}
 end
 
-function _buffered_simplify(buffer, x::Rational)
-    operate_to!(buffer, gcd, x.num, x.den)
-    if !isone(buffer)
-        operate!(div, x.num, buffer)
-        operate!(div, x.den, buffer)
-    end
+function buffer_for(
+    ::Union{typeof(+),typeof(-)},
+    ::Type{Rational{S}},
+    ::Type{Rational{T}},
+) where {S,T}
+    U = promote_operation(gcd, S, T)
+    return zero(U), zero(U), zero(U)
+end
+
+function buffered_operate_to!(
+    buffer::Tuple,
+    output::Rational,
+    op::Union{typeof(-),typeof(+)},
+    x::Rational,
+    y::Rational,
+)
+    _buffered_divgcd(buffer[1], x.den, buffer[2], y.den, buffer[3])
+    # TODO: Use `checked_mul` and `checked_sub` like in Base
+    operate_to!(output.num, *, x.num, buffer[3])
+    buffered_operate!(
+        buffer[2],
+        add_sub_mul_op(op),
+        output.num,
+        y.num,
+        buffer[2],
+    )
+    operate_to!(output.den, *, x.den, buffer[3])
+    _buffered_simplify(buffer[1], output)
+    return output
+end
+
+function operate_to!(
+    output::Rational,
+    op::Union{typeof(+),typeof(-)},
+    x::Rational,
+    y::Rational,
+)
+    return buffered_operate_to!(
+        buffer_for(op, typeof(x), typeof(y)),
+        output,
+        op,
+        x,
+        y,
+    )
 end
 
 function operate_to!(output::Rational, ::typeof(+), x::Rational, y::Rational)
@@ -56,27 +112,6 @@ function operate_to!(output::Rational, ::typeof(+), x::Rational, y::Rational)
     # TODO: Use `checked_mul` and `checked_add` like in Base
     operate_to!(output.num, *, x.num, yd)
     operate!(add_mul, output.num, y.num, xd)
-    operate_to!(output.den, *, x.den, yd)
-    # Reuse `xd` as it is a local copy created by this method
-    _buffered_simplify(xd, output)
-    return output
-end
-
-# -
-
-function promote_operation(
-    ::typeof(-),
-    ::Type{Rational{S}},
-    ::Type{Rational{T}},
-) where {S,T}
-    return Rational{promote_sum_mul(S, T)}
-end
-
-function operate_to!(output::Rational, ::typeof(-), x::Rational, y::Rational)
-    xd, yd = Base.divgcd(promote(x.den, y.den)...)
-    # TODO: Use `checked_mul` and `checked_sub` like in Base
[…]
