### 2023-02-14 c023e92
Add operate for div, gcd, lcm
