### 2023-02-14 15986a5
Add operate for div, gcd, lcm (#196)
* Add operate for div, gcd, lcm

* Fix format
