+++ src/interface.jl
@@ -191,7 +191,14 @@ function operate end
 operate(::typeof(-), x) = -x
 
 function operate(
-    op::Union{typeof(+),typeof(*),AddSubMul,typeof(add_dot)},
+    op::Union{
+        typeof(+),
+        typeof(*),
+        AddSubMul,
+        typeof(add_dot),
+        typeof(gcd),
+        typeof(lcm),
+    },
     x,
     y,
     args::Vararg{Any,N},
@@ -199,13 +206,15 @@ function operate(
     return op(x, y, args...)
 end
 
-operate(op::Union{typeof(-),typeof(/)}, x, y) = op(x, y)
+operate(op::Union{typeof(-),typeof(/),typeof(div)}, x, y) = op(x, y)
 
 operate(::typeof(convert), ::Type{T}, x) where {T} = convert(T, x)
 
 operate(::typeof(convert), ::Type{T}, x::T) where {T} = copy_if_mutable(x)
 
-operate(::Union{typeof(+),typeof(*)}, x) = copy_if_mutable(x)
+function operate(::Union{typeof(+),typeof(*),typeof(gcd),typeof(lcm)}, x)
+    return copy_if_mutable(x)
+end
 
 # We could only give `typeof(x)` to `zero` and `one` to be sure that modifying
 # the returned object cannot alter `x` but for some objects, `one` and `zero`
