### 2023-02-14 9596e66
Support x[;kwargs...] syntax in rewrite macro
