+++ test/rewrite.jl
@@ -167,5 +167,5 @@ Base.getindex(x::_KwargRef; i) = x.data[i]
 
 @testset "test_rewrite_kw_in_ref" begin
     x = _KwargRef(Dict(i => i + 1 for i in 2:4))
-    @test MA.@rewrite(sum(x[i=j] for j in 2:4)) == 12
+    @test MA.@rewrite(sum(x[i = j] for j in 2:4)) == 12
 end
