+++ versions.js
@@ -8,5 +8,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.2.2";
+var DOCUMENTER_NEWEST = "v1.2.3";
 var DOCUMENTER_STABLE = "stable";
