### 2023-03-20 a02ba9a
Add similar_array_type for Linearalgebra.Diagonal
