+++ src/implementations/LinearAlgebra.jl
@@ -155,6 +155,13 @@ function similar_array_type(
     return LinearAlgebra.Symmetric{S,similar_array_type(MT, S)}
 end
 
+function similar_array_type(
+    ::Type{LinearAlgebra.Diagonal{T,VT}},
+    ::Type{S},
+) where {S,T,VT<:AbstractVector{T}}
+    return LinearAlgebra.Diagonal{S,Vector{S}}
+end
+
 similar_array_type(::Type{Array{T,N}}, ::Type{S}) where {S,T,N} = Array{S,N}
 
 similar_array_type(::Type{BitArray{N}}, ::Type{S}) where {S,N} = Array{S,N}
