### 2023-03-22 6f6a71f
Fix init in summation rewrite
