### 2023-03-22 86a502e
Fix init in summation rewrite (#207)
