### 2023-04-19 2d2851e
BigFloat dot: add a promote_operation method
Seems like MA requires these?
