+++ test/runtests.jl
@@ -22,6 +22,9 @@ end
 @testset "BigFloat negation and absolute value" begin
     include("bigfloat_neg_abs.jl")
 end
+@testset "BigFloat fma and muladd" begin
+    include("bigfloat_fma.jl")
+end
 @testset "BigFloat dot" begin
     include("bigfloat_dot.jl")
 end
