### 2023-04-20 710e5b5
in tests, use MA.copy_if_mutable for copying
... instead of reimplementing it ad-hoc
