### 2023-04-21 3d5ae74
in tests, use MA.copy_if_mutable for copying (#215)
... instead of reimplementing it ad-hoc
