### 2023-04-26 677e5e7
add evalpoly (#214)
```
julia> setprecision(512)
512

julia> @benchmark evalpoly(x, coefs) setup=(x = rand(BigFloat); coefs = rand(BigFloat, 10);)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.247 μs … 276.376 μs  ┊ GC (min … max): 0.00% … 96.44%
 Time  (median):     1.487 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.611 μs ±   3.876 μs  ┊ GC (mean ± σ):  3.30% ±  1.37%

         ▁▃▄▇█▇▃
  ▂▂▂▃▄▆████████▇▅▄▃▄▄▄▅▅▅▅▅▅▄▅▄▄▃▃▃▃▃▂▂▂▂▂▂▂▂▂▁▁▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  1.25 μs         Histogram: frequency by time        2.41 μs <

 Memory estimate: 1.20 KiB,
[…]
