### 2023-04-28 cb5b915
use `StridedMaybeAdjOrTransMat` instead of defining it
