### 2023-05-02 211e34e
use `StridedMaybeAdjOrTransMat` instead of defining it (#217)
