### 2023-09-01 72dbd20
Fix rewrite of +(x, *(y...)) into add_mul
