### 2023-09-01 01cae64
Clarify docstring of rewrite macro (#226)
