+++ src/rewrite.jl
@@ -5,7 +5,7 @@
 # one at http://mozilla.org/MPL/2.0/.
 
 """
-    @rewrite(expr; move_factors_into_sums = false)
+    @rewrite(expr, move_factors_into_sums = false)
 
 Return the value of `expr`, exploiting the mutability of the temporary
 expressions created for the computation of the result.
@@ -13,6 +13,10 @@ expressions created for the computation of the result.
 If you have an `Expr` as input, use [`rewrite_and_return`](@ref) instead.
 
 See [`rewrite`](@ref) for an explanation of the keyword argument.
+
+!!! info
+    Passing `move_factors_into_sums` after a `;` is not supported. Use a `,`
+    instead.
 """
 macro rewrite(args...)
     @assert 1 <= length(args) <= 2
