### 2023-09-06 774bc08
Fix rewrite of +(x, *(y...)) into add_mul (#225)
* Fix rewrite of +(x, *(y...)) into add_mul

* Add test

* Add more tests
