### 2023-09-20 2101304
Fix bug rewriting sum with dims kwarg
