+++ .github/workflows/format_check.yml
@@ -13,13 +13,13 @@ jobs:
       - uses: julia-actions/setup-julia@latest
         with:
           version: '1'
-      - uses: actions/checkout@v1
+      - uses: actions/checkout@v3
       - name: Format check
         shell: julia --color=yes {0}
         run: |
           using Pkg
           # If you update the version, also update the style guide docs.
-          Pkg.add(PackageSpec(name="JuliaFormatter", version="0.20.4"))
+          Pkg.add(PackageSpec(name="JuliaFormatter", version="1"))
           using JuliaFormatter
           format(".", verbose=true)
           out = String(read(Cmd(`git diff`)))
