+++ .github/workflows/aqua.yml
@@ -12,7 +12,7 @@ jobs:
       - uses: julia-actions/setup-julia@latest
         with:
           version: '1'
-      - uses: actions/checkout@v1
+      - uses: actions/checkout@v3
       - name: Aqua
         shell: julia --color=yes {0}
         run: |
