### 2023-09-20 37bb9bf
Fix bug rewriting sum with dims kwarg (#229)
