+++ versions.js
@@ -9,5 +9,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.3.2";
+var DOCUMENTER_NEWEST = "v1.3.3";
 var DOCUMENTER_STABLE = "stable";
