### 2023-10-06 58bb575
Minor refactor with _is_sum
