### 2023-10-11 1c70967
Add support for BitVector and BitArray in operate
