### 2023-10-12 1470f20
Update test/interface.jl
Co-authored-by: Mathieu Besançon <mathieu.besancon@gmail.com>
