### 2023-10-12 6e6a854
Add support for BitVector and BitArray in operate (#233)
