### 2023-11-05 11acff5
add dispatch for dot between Symmetric and Hermitian
