### 2023-11-07 #93 @blegat
@lucaferranti I have tried `latexmk -bibtex -pdf paper.tex`  suggested by the [author guide](https://juliacon.github.io/proceedings-guide/author/) and it works locally. Is there any way to see the log of whedon's build ?
