### 2023-11-07 #93 @lucaferranti
@blegat unfortunately logs are not accessible. This and a lot of other latex issues will be fixed once we migrate to the new `editorial-bot` (still WIP, but should happen within a few weeks).

Do notice that here we use a fairly old version of `latexmk` (`17 Jan. 2018, version: 4.55.`), so the issue might be that when updating the paper you used some latex features not supported by the older version?
