### 2023-11-24 #93 @schillic
#### Summary

The paper presents a well thought through Julia package with an elegant API. The paper itself is very well written with nice motivating examples, which makes it easy to follow for a reader with basic Julia knowledge. I have a few requests to fix some code blocks and update the references, thus a *Minor modification request*. There are quite some further suggestions for improvement, but none of them is critical for publication. So @blegat I leave it to your discretion to incorporate them. Just give @lucaferranti a sign when the paper is ready from your side. I can also take a look again if you prefer. Happy to hopefully see this paper published soon!

#### Please fix

- `add_mul!(...)` should be `add_mul!!(...)` (2x in code blocks on pp. 4 & 5).
- The third code block on p. 5 starting with `buf = buffer_for(` does not execute for me (see the error below)
  - Should the `typeof`s be `eltype`s as in the first block?
  - The call to `buffered_operate!` also fails. I guess it should be the same as in the second block, just prepended by `@allocated`?
```julia
ERROR: MethodError: no method matching buffer_for(::typeof(add_mul), ::Type{Vector{BigInt}}, ::Type{Matrix{BigInt}}, ::Type{Vector{BigInt}})

Closest candidates are:
  buffer_for(::typeof(add_mul), ::Type{BigInt}...)
```
- The fourth code block on p. 5 starting with `using Polynomials` does not work because `m` is undefined. Please add the value you used so that it works and the results are reproducible.
- In the fifth code block on p. 5 (last code block in the paper), the last two commands do not work.
- References:
  - [1],[8],[9] can probably update the version
  - [5] has a published version: https://dblp.org/rec/journals/cca/GowdaMCGSER21.html?view=bibtex
  - [6] is published now: https://dblp.org/rec/journals/informs/LegatDGL22.html?view=bibtex
  - [7] should use proper capitalization

#### Checklist comments

- A statement of need: While there is no dedicated section, the purpose of the package is clearly communicated.
- Installation instructions: While there are no instructions, the standard `Pkg` installation is smooth. I consider this sufficient.
- Community guidelines: While there are no guidelines in the README or documentation, I believe jump-dev is well known in the community and is easy to reach out to. Nevertheless, maybe you want to include a link in the README.

#### Suggestions

- I guess you want to update your affiliation.
- I suggest to avoid the page breaks in code blocks. You can wrap the `jllisting`s in a `minipage` environment, possibly with some negative `\vspace` at the beginning (I tried it and it works).
- I suggest to define a hyphenation rule for "MutableArithmetics" via `\hyphenation{Mutable-Arithmetics}`.
- abstract: "talk" -> "package" (or "paper")
- p. 1:
  - "when summing the elements of an array" - I did not understand this example at this point (why would you mutate the array here?). I find the other example (array multiplication) sufficiently convincing, so maybe the first example can just be dropped here; maybe the multiplication example can be generalized to array arithmetic (because I do not think there is anything special about multiplication).
  - Footnote 1 is missing a period in the end.
  - "multiple precision numbers" -> I would write "multiple-precision numbers" for clarity.
  - In the enumeration with (1) and (2) I would start with "MutableArithmetics" instead of "it".
  - "e.g" -> "e.g.,"
  - When recalling (1) and (2) via `\ref`, I would use `\eqref` to get the parentheses (2x).
  - "on top of the MutableArithmetics API" - I am not sure what this is supposed to mean. Maybe it can simply be removed?
  - "Julia sum function" -> "Julia's sum function"
  - "cases where" -> "cases, where"
  - "a linear combinations" -> "linear combinations"
- p. 2:
  - code blocks: Maybe indent lines that continue from the previous line (e.g., by two spaces).
  - "would fallback" -> "would fall back"
  - "writing a code that should" -> I would say "writing code to" or "writing a function to".
  - "the rational Julia type" - I would use green "Rational" as you do for BigInt etc. to avoid an ambiguity here. I would also flip to "the Julia Rational type" or even drop "Julia" altogether.
  - I am not sure about referring back to footnote 1 two more times here. You already clarified what you mean by "mutate", so I do not see the need to repeat that. But I guess this is a matter of taste.
  - "of rational number introduces" -> "of rational numbers introduced"
  - "may mutate [...] should mutate" - I would emphasize (`\emph`) "may" and "should".
  - In the last code block, `function Base.:*(A::Matrix{S}, b::Vector{T})` leads to an error because `S` and `T` are undefined (not very relevant because the whole code block does not compile (on purpose) anyway).
- p. 3:
  - "mutate^1 args[1]" -> "mutates^1 args[1]"
  - "operations so" -> "operations, so"
  - "each pairs" -> "each pair"
  - "value of the result need" -> "value of the result needs"
  - "the non-mutable type" -> "the non-mutable types"
  - The first method in the last code block could use the short notation (`promote_operation(...) = Foo`). I mention this because you used it elsewhere.
- p. 4:
  - The second code block in Section 4 is missing the squaring (`^2`).
  - The output of BenchmarkTools has looks quite different (see below). I do not suggest to change the output in the paper and just wanted to give a heads up.
  - "multiple occurence" -> "multiple occurrences"
  - The first method in the last code block could again use the short notation.
  - "there is only 48 bytes" -> "there are only 48 bytes"
  - "this is the allocation" -> "which is the allocation"
  - The last code block has a `where {N}` that is not needed.
```julia
BenchmarkTools.Trial: 334 samples with 1 evaluation.
 Range (min … max):   7.181 ms … 105.259 ms  ┊ GC (min … max):  0.00% … 53.51%
 Time  (median):      7.376 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):   15.003 ms ±  24.655 ms  ┊ GC (mean ± σ):  28.72% ± 14.81%

  █
  ██▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇█▆▅ ▅
  7.18 ms       Histogram: log(frequency) by time      99.1 ms <

 Memory estimate: 3.67 MiB, allocs estimate: 198166.
```
- p. 5:
  - "higher level operations" -> "higher-level operations"
  - "one of the motivation" -> "one of the motivations" or "one motivation"
  - "a interface" -> "an interface"
  - "expressions but" -> "expressions, but"
  - "of implementation of" -> "of implementing" (sounds better)
- acknowledgments: The last paragraph should be updated (there is only one author).
