### 2023-11-24 #93 @schillic
@lucaferranti: I ticked off the checklist. I left "Functionality" open to address the fixes to the code blocks as indicated in my review.
