### 2023-11-26 aa79211
`copy` for `BigInt`, `BigFloat` and `Rational`
Fixes #163.
