+++ src/implementations/BigInt.jl
@@ -13,6 +13,17 @@ mutability(::Type{BigInt}) = IsMutable()
 # https://github.com/JuliaLang/julia/blob/7d41d1eb610cad490cbaece8887f9bbd2a775021/base/gmp.jl#L772
 mutable_copy(x::BigInt) = Base.GMP.MPZ.set(x)
 
+# copy
+
+promote_operation(::typeof(copy), ::Type{BigInt}) = BigInt
+
+function operate_to!(out::BigInt, ::typeof(copy), in::BigInt)
+    Base.GMP.MPZ.set!(out, in)
+    return out
+end
+
+operate!(::typeof(copy), x::BigInt) = x
+
 # zero
 
 promote_operation(::typeof(zero), ::Type{BigInt}) = BigInt
