### 2023-11-26 f520983
add an `operate` method for `abs`
Fixes #241
