+++ test/interface.jl
@@ -97,7 +97,7 @@ end
 @testset "operate" begin
     @testset "$T" for T in (Int, BigInt, Rational{Int})
         x = T(7)
-        @testset "1-ary $op" for op in [+, *, gcd, lcm]
+        @testset "1-ary $op" for op in [+, *, gcd, lcm, abs]
             a = op(x)
             b = MA.operate(op, x)
             @test a == b
