+++ Project.toml
@@ -9,7 +9,11 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
+LinearAlgebra = "<0.0.1, 1.6"
 OffsetArrays = "1"
+Random = "<0.0.1, 1.6"
+SparseArrays = "<0.0.1, 1.6"
+Test = "<0.0.1, 1.6"
 julia = "1.6"
 
 [extras]
