### 2023-11-27 0719307
implement `abs` for `BigInt` and `Rational`
Tests coming in linked PRs.
