### 2023-11-27 ac7a8a9
Fix incorrect `operate` test set (#242)
