### 2023-11-27 4cc20ad
Add an `operate` method for `abs` (#243)
