### 2023-11-27 3a6c406
dispatch.jl: simplify overloads for SparseMatrix * AbstractArray (#181)
