+++ src/implementations/LinearAlgebra.jl
@@ -163,7 +163,6 @@ function similar_array_type(
 end
 
 similar_array_type(::Type{<:AbstractVector}, ::Type{T}) where {T} = Vector{T}
-    
 similar_array_type(::Type{Array{T,N}}, ::Type{S}) where {S,T,N} = Array{S,N}
 
 similar_array_type(::Type{BitArray{N}}, ::Type{S}) where {S,N} = Array{S,N}
