### 2023-11-28 c838f91
Add similar_array_type for LinearAlgebra.Diagonal (#205)
