### 2023-11-28 28125f4
`copy` for `BigInt`, `BigFloat` and `Rational` (#238)
Fixes #163.
