### 2023-11-28 444331d
Implement BigFloat negation and abs without ccall (#194)
