### 2023-11-30 229103a
Add dispatch for LinearAlgebra.dot for Symmetric and Hermitian matrices (#248)
Co-authored-by: araujoms <maltusan@gmail.com>
