### 2023-11-30 9b58a40
Add some methods for unary `+`, `-`, `*`, `abs` (#245)
This is the first time that `abs` is implemented for `BigInt` and
`Rational`.

Fixes #237.
