+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.4",
   "v1.3",
   "v1.2",
   "v1.1",
@@ -9,5 +10,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.3.3";
+var DOCUMENTER_NEWEST = "v1.4.0";
 var DOCUMENTER_STABLE = "stable";
