### 2023-12-21 5d1541f
Fix implementation of mul! for AbstractMatrix and AbstractVector
