### 2023-12-27 c096dd7
[docs] fix docstring of rewrite macro (#254)
