### 2024-01-09 85f5ff2
Fix implementation of mul! for AbstractMatrix and AbstractVector (#252)
