### 2024-02-13 f75f3ad
Fix Vector*Transpose{Vector}
