### 2024-02-14 787bfe2
Fix a test for nightly (#258)
