### 2024-02-14 22e0bac
Fix multiplying Symmetric and Hermitian matrices
