### 2024-02-14 3a1ca9b
Fix Vector*Transpose{Vector} (#257)
