### 2024-02-14 38b9654
Fix promotion of outer product
