### 2024-02-14 70b043a
Fix promotion of outer product (#263)
* Fix promotion of outer product

* Update matmul.jl

---------

Co-authored-by: Oscar Dowson <odow@users.noreply.github.com>
