+++ versions.js
@@ -10,5 +10,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.4.0";
+var DOCUMENTER_NEWEST = "v1.4.1";
 var DOCUMENTER_STABLE = "stable";
