### 2024-02-29 e8aedaa
Fix product of Complex with Hermitian
