+++ test/dispatch.jl
@@ -55,3 +55,11 @@ end
     @test all(MA.isequal_canonical.(2 * LinearAlgebra.Hermitian(A, :L), D))
     @test all(MA.isequal_canonical.(2 * LinearAlgebra.Hermitian(A, :U), D))
 end
+
+@testset "*(::Complex, ::Hermitian)" begin
+    A = [1 2; 2 3]
+    B = LinearAlgebra.Hermitian(DummyBigInt.(A))
+    C = 2im * A
+    @test 2im * B == C
+    @test C isa Matrix{Complex{BigInt}}
+end
