### 2024-02-29 69fbcaa
Fix product of Complex with Hermitian (#265)
