+++ src/dispatch.jl
@@ -484,7 +484,10 @@ function Base.:*(α::Number, A::LinearAlgebra.Symmetric{<:AbstractMutable})
     return LinearAlgebra.Symmetric(B, c)
 end
 
-function Base.:*(α::AbstractMutable, A::LinearAlgebra.Symmetric{<:AbstractMutable})
+function Base.:*(
+    α::AbstractMutable,
+    A::LinearAlgebra.Symmetric{<:AbstractMutable},
+)
     c = LinearAlgebra.sym_uplo(A.uplo)
     B = c == :U ? _mult_upper(α, A) : _mult_lower(α, A)
     return LinearAlgebra.Symmetric(B, c)
