### 2024-03-02 97cf4ce
Update src/dispatch.jl
Co-authored-by: Mateus Araújo <maltusan@gmail.com>
