### 2024-03-03 a7b10f2
Update src/dispatch.jl
Co-authored-by: Mateus Araújo <maltusan@gmail.com>
