### 2024-03-04 dfce02a
Fix *(::AbstractMutable, ::Symmetric) (#268)
Co-authored-by: Mateus Araújo <maltusan@gmail.com>
