### 2024-03-07 332b103
Make promote_operation work for division
