+++ paper/paper.tex
@@ -1,12 +1,15 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
 
+\usepackage{amsmath}
 \usepackage{cleveref}
 \crefformat{footnote}{#2\footnotemark[#1]#3}
 
 \newcommand{\ma}{MutableArithmetics}
 
 \addlitjlmacros{@rewrite}{@rewrite}{8}
+
+\hyphenation{Mutable-Arithmetics}
 
 \begin{document}
 
@@ -19,7 +22,7 @@
 Arithmetic operations defined in Julia do not modify their arguments.
 However, in many situations, a variable represents an accumulator that can be modified in-place to contain the result, e.g., when summing the elements of an array.
 Moreover, for types that support mutation, mutating the value may have a significant performance benefit over creating a new instance.
-This talk presents an interface that allows algorithms to exploit mutability in arithmetic operations in a generic manner.
+This paper presents an interface that allows algorithms to exploit mutability in arithmetic operations in a generic manner.
 
 \end{abstract}
 
@@ -28,10 +31,10 @@ This talk presents an interface that allows algorithms to exploit mutability in 
 Julia enables developers to write generic algorithms that work with arbitrary number types, as long as the types implement the needed operations
 such as \lstinline|+|, \lstinline|*|, \lstinline|-|, \lstinline|zero|, \lstinline|one|, ...
 The implementations of these arithmetic operations in Julia do not modify their arguments. Instead, they return a new instance of the type as the result.
-However, in many situations, a variable represents an accumulator that can be mutated\footnote{\label{foot:mutate}In this paper, the terminology ``mutate $x$ to $y$'' means modifying $x$ in-place in such a way that its value after the modification is equal to $y$} to the result, e.g.,
-when summing the elements of an array or when implementing array multiplication.
+However, in many situations, a variable represents an accumulator that can be mutated\footnote{\label{foot:mutate}In this paper, the terminology ``mutate $x$ to $y$'' means modifying $x$ in-place in such a way that its value after the modification is equal to $y$.} to the result, e.g.,
+when summing the elements of an array of \lstinline|BigInt| or when implementing array multiplication.
 Moreover, for types that support mutation, mutating the value may have a significant performance benefit over creating a new instance.
-Examples of types that implement arithmetic operations and support mutation include \lstinline|Array|s, multiple precision numbers, JuMP~\cite{dunning2017jump} expressions, MathOptInterface (MOI)~\cite{legat2021mathoptinterface} functions, and polynomials (univariate~\cite{verzani2021polynomials} or multivariate~\cite{legat2021multivariatepolynomials}).
+Examples of types that implement arithmetic operations and support mutation include \lstinline|Array|s, multiple-precision numbers, JuMP~\cite{dunning2017jump} expressions, MathOptInterface (MOI)~\cite{legat2021mathoptinterface} functions, and polynomials (univariate~\cite{verzani2021polynomials} or multivariate~\cite{legat2021multivariatepolynomials}).
 
 This paper introduces an interface called \ma{}.
 It allows mutable types to implement an arithmetic exploiting their mutability, and for algorithms to
@@ -40,18 +43,19 @@ Moreover, it provides the following additional features:
 \begin{enumerate}
   \item
     \label{item:reimplement}
-    it re-implements part of the Julia standard library on top of the API to allow mutable types to use a more efficient version than the default one.
[…]
