+++ paper/header.tex
@@ -3,7 +3,7 @@
 \title{MutableArithmetics: An API for mutable operations}
 
 \author[1]{Benoît Legat}
-\affil[1]{Massachusetts Institute of Technology}
+\affil[1]{KU Leuven}
 
 \keywords{Julia, Optimization, Performance, Interface}
 
