+++ .vscode/ltex.dictionary.en-US.txt
@@ -1,0 +1,11 @@
+LinearAlgebra
+BenchmarkTools
+@benchmark
+mul
+BigInt
+SumOfSquares
+JuMP
+MutableArithmetics
+Dowson
+Lubin
+MultivariatePolynomials
