+++ paper/paper.yml
@@ -9,7 +9,7 @@ authors:
     orcid: 0000-0002-5921-368X
     affiliation: 1
 affiliations:
-  - name: Massachusetts Institute of Technology
+  - name: KU Leuven
     index: 1
 date: 29 July 2021
 bibliography: ref.bib
