+++ paper/paper.tex
@@ -65,15 +65,7 @@ The section is organized into four subsections that describe the need of four ke
 \label{sec:may_mutate}
 Consider the task of summing the elements of a vector.
 By default, Julia's \lstinline|sum| function will compute the sum with a code equivalent to the following:
-\begin{jllisting}
-function sum(x::Vector)
-    acc = zero(eltype(x))
-    for el in x
-        acc = acc + el
-    end
-    return acc
-end
-\end{jllisting}
+\jlinputlisting{code/sum.jl}
 If the type of the elements of \lstinline|x| is \lstinline|BigInt|, it is more efficient to replace the line
 \lstinline|acc = acc + el| by the line
 \lstinline|Base.GMP.MPZ.add!(acc, el)|.
@@ -89,17 +81,7 @@ We now consider a mutable element type for which exploiting mutability affects t
 Consider a type \lstinline|SymbolicVariable| representing a symbolic variable
 and the following types representing linear combinations of these variables with coefficients of type \lstinline|T|.
 This example encapsulates for instance JuMP affine expressions~\cite{dunning2017jump}, MOI affine functions~\cite{legat2021mathoptinterface}, polynomials (univariate~\cite{verzani2021polynomials} or multivariate~\cite{legat2021multivariatepolynomials}) or symbolic sums~\cite{gowda2021high}.
-\begin{jllisting}
-struct Term{T}
-    coef::T
-    sym::SymbolicVariable
-end
-struct Sum{T}
-    terms::Vector{Term{T}}
-end
-Base.:+(s::Sum, t::Term) = Sum(push!(copy(s.terms), t))
-Base.zero(::Type{Term{T}}) where {T} = Sum(Term{T}[])
-\end{jllisting}
+\jlinputlisting{code/term.jl}
 Calling \lstinline|sum| on a vector of $n$ \lstinline|Term{T}| has a time complexity $\Theta(n^2)$.
 Indeed, when calling \lstinline|acc + el| where \lstinline|acc| contains the sum of the first \lstinline|k| terms and \lstinline|el| is the $(k+1)$th term,
 the result cannot mutate \lstinline|acc.terms| and the copy of \lstinline|acc.terms| has time complexity $\Theta(k)$.
@@ -111,21 +93,7 @@ By default, \lstinline|add!!| would fall back to calling \lstinline|+|
 so that a method calling \lstinline|add!!| would both exploit the mutability
 of mutable types but would also work for non-mutable types.
 For our example, an implementation could be:
-\begin{jllisting}
-function sum(x)
-    acc = zero(eltype(x))
-    for el in x
-        acc = add!!(acc, el)
-    end
-    return acc
-end
-add!!(a, b) = a + b # default fallback
-add!!(a::BigInt, b::BigInt) = Base.GMP.MPZ.add!(a, b)
-function add!!(s::Sum, t::Term)
-    push!(s.terms, t)
-    return s
-end
-\end{jllisting}
+\jlinputlisting{code/termsum.jl}
 Note that the time complexity of the sum of $n$ \lstinline|Term| is now $\Theta(n)$.
 
 Julia implements a specialized method for computing the sum of \lstinline|BigInt|s that uses \lstinline|Base.GMP.MPZ.add!|.
@@ -144,12 +112,7 @@ an API that allows an implementation to silently returns the
 result without modifying the first argument is not appropriate.
 
 To motivate this, consider the \lstinline|Rational| Julia type:
-\begin{jllisting}
-struct Rational{T}
-    num::T
-    den::T
-end
-\end{jllisting}
+\jlinputlisting{code/rational.jl}
 Suppose we want to mutate\cref{foot:mutate} some rational \lstinline|a::Rational| to the
[…]
