+++ paper/header.tex
@@ -9,8 +9,7 @@
 
 \hypersetup{
 pdftitle = {MutableArithmetics: An API for mutable operations},
-pdfsubject = {JuliaCon 2019 Proceedings},
+pdfsubject = {JuliaCon 2021 Proceedings},
 pdfauthor = {Benoît Legat},
 pdfkeywords = {Julia, Optimization, Performance, Interface},
 }
-
