### 2024-03-21 ddfef20
Test that mutability runs with no error
