### 2024-03-23 #93 @lucaferranti
jippiie we have the new editorialbot and paper compilation works :tada:

@blegat @schillic can you remind me the status of the review?
