### 2024-03-24 2481775
Fix mutability check for array multiplication
