+++ test/interface.jl
@@ -178,3 +178,12 @@ end
     @test MA.operate_to!!(T(6), abs, T(7)) == 7
     @test MA.operate_to!!(T(6), abs, T(-7)) == 7
 end
+
+@testset "Error-free mutability (issue #240)" begin
+    for op in (+, -, *, /, div)
+        for T in
+            (Float64, BigFloat, Int, BigInt, Rational{Int}, Rational{BigInt})
+            @test_nowarn MA.mutability(T, op, T, T) # should run without error
+        end
+    end
+end
