+++ paper/code/verify.jl
@@ -3,7 +3,7 @@ using MutableArithmetics
 struct SymbolicVariable end
 for file in readdir(@__DIR__)
     path = joinpath(@__DIR__, file)
-    if path != (@__FILE__) && file != "mul.jl"
+    if path != (@__FILE__)
         include(file)
     end
 end
