+++ paper/ref.bib
@@ -50,18 +50,6 @@
     doi={10.1287/ijoc.2021.1067},
     publisher={INFORMS}
 }
-@software{legat2021multivariatepolynomials,
-  author       = {Benoît Legat and
-                  Sascha Timme and
-                  Robin Deits},
-  title        = {{JuliaAlgebra/MultivariatePolynomials.jl: v0.5.4}},
-  month        = jan,
-  year         = 2024,
-  publisher    = {Zenodo},
-  version      = {v0.5.4},
-  doi          = {10.5281/zenodo.10468722},
-  url          = {https://zenodo.org/doi/10.5281/zenodo.10468722}
-}
 
 @software{verzani2021polynomials,
   author       = {John Verzani and
@@ -136,4 +124,14 @@
   booktitle = {JuliaCon},
   year      = {2019},
   url       = {https://pretalx.com/juliacon2019/talk/QZBKAU/},
+}
+
+@Conference{legat2023multivariate,
+  author       = {Legat, Beno{\^\i}t},
+  title        = {Multivariate polynomials in Julia},
+  month        = jul,
+  year         = 2022,
+  booktitle    = {JuliaCon},
+  year         = {2022},
+  url          = {https://pretalx.com/juliacon-2022/talk/TRFSJY/},
 }
