+++ paper/ref.bib
@@ -128,7 +128,7 @@
 
 @Conference{legat2023multivariate,
   author       = {Legat, Beno{\^\i}t},
-  title        = {Multivariate polynomials in Julia},
+  title        = {Multivariate polynomials in {Julia}},
   month        = jul,
   year         = 2022,
   booktitle    = {JuliaCon},
