+++ paper/code/rational.jl
@@ -1,0 +1,4 @@
+struct Rational{T}
+    num::T
+    den::T
+end
