+++ paper/.vscode/ltex.dictionary.en-US.txt
@@ -1,0 +1,19 @@
+LinearAlgebra
+BenchmarkTools
+@benchmark
+mul
+BigInt
+SumOfSquares
+JuMP
+MutableArithmetics
+Dowson
+Lubin
+MultivariatePolynomials
+jl
+gcd
+num
+BangBang
+MathOptInterface
+@rewrite
+JuMP-specific
+MPZ
