+++ paper/ref.bib
@@ -39,25 +39,16 @@
   publisher = {SIAM},
 }
 
-@Article{legat2021mathoptinterface,
-  author  = {Legat, Beno\^{\i}t and Dowson, Oscar and Garcia, Joaquim Dias and Lubin, Miles},
-  journal = {INFORMS Journal on Computing},
-  title   = {{MathOptInterface}: a data structure for mathematical optimization problems},
-  year    = {2021},
-  note    = {In press},
-}
-
-@software{legat2021multivariatepolynomials,
-  author       = {Benoît Legat and
-                  Sascha Timme and
-                  Robin Deits},
-  title        = {{JuliaAlgebra/MultivariatePolynomials.jl: v0.3.18}},
-  month        = jul,
-  year         = 2021,
-  publisher    = {Zenodo},
-  version      = {v0.3.18},
-  doi          = {10.5281/zenodo.5083796},
-  url          = {https://doi.org/10.5281/zenodo.5083796}
+@article{legat2021mathoptinterface,
+    title={{MathOptInterface}: a data structure for mathematical optimization problems},
+    author={Legat, Beno{\^\i}t and Dowson, Oscar and Dias Garcia, Joaquim and Lubin, Miles},
+    journal={INFORMS Journal on Computing},
+    year={2021},
+    volume={34},
+    number={2},
+    pages={672--689},
+    doi={10.1287/ijoc.2021.1067},
+    publisher={INFORMS}
 }
 
 @software{verzani2021polynomials,
@@ -65,23 +56,37 @@
                   Miles Lucas and
                   Zdeněk Hurák and
                   Jameson Nash},
-  title        = {{JuliaMath/Polynomials.jl: v2.0.14}},
-  version      = {v2.0.14},
+  title        = {{JuliaMath/Polynomials.jl: v4.0.5}},
+  version      = {v4.0.5},
   url          = {https://github.com/JuliaMath/Polynomials.jl}
 }
 
 @software{takafumi2021bangbang,
   author       = {Takafumi Arakaki},
-  title        = {{JuliaFolds/BangBang.jl: v0.3.31}},
-  version      = {v0.3.31},
-  url          = {https://github.com/JuliaFolds/BangBang.jl}
+  title        = {{JuliaFolds2/BangBang.jl: v0.4.1}},
+  version      = {v0.4.1},
+  url          = {https://github.com/JuliaFolds2/BangBang.jl}
 }
 
 @article{gowda2021high,
-  title={High-performance symbolic-numerics via multiple dispatch},
-  author={Gowda, Shashi and Ma, Yingbo and Cheli, Alessandro and Gwozdz, Maja and Shah, Viral B and Edelman, Alan and Rackauckas, Christopher},
-  journal={arXiv preprint arXiv:2105.03949},
-  year={2021}
+  author       = {Shashi Gowda and
+                  Yingbo Ma and
+                  Alessandro Cheli and
+                  Maja Gw{\'{o}}zdz and
+                  Viral B. Shah and
+                  Alan Edelman and
+                  Christopher Rackauckas},
+  title        = {High-performance symbolic-numerics via multiple dispatch},
+  journal      = {{ACM} Commun. Comput. Algebra},
+  volume       = {55},
+  number       = {3},
+  pages        = {92--96},
+  year         = {2021},
+  url          = {https://doi.org/10.1145/3511528.3511535},
+  doi          = {10.1145/3511528.3511535},
+  timestamp    = {Sat, 28 Jan 2023 18:54:19 +0100},
+  biburl       = {https://dblp.org/rec/journals/cca/GowdaMCGSER21.bib},
[…]
