+++ paper/header.tex
@@ -3,14 +3,13 @@
 \title{MutableArithmetics: An API for mutable operations}
 
 \author[1]{Benoît Legat}
-\affil[1]{Massachusetts Institute of Technology}
+\affil[1]{KU Leuven}
 
 \keywords{Julia, Optimization, Performance, Interface}
 
 \hypersetup{
 pdftitle = {MutableArithmetics: An API for mutable operations},
-pdfsubject = {JuliaCon 2019 Proceedings},
+pdfsubject = {JuliaCon 2021 Proceedings},
 pdfauthor = {Benoît Legat},
 pdfkeywords = {Julia, Optimization, Performance, Interface},
 }
-
