### 2024-03-26 69b1c51
Fix mutability check for array multiplication (#272)
