### 2024-04-05 7f10e04
Allow sum of arrays and product with scalar
