### 2024-04-08 #93 @schillic
Thanks for the new version. There are still some issues with the code in the paper. I think they should be fixed, at least the one marked with :one:. The ones marked with :two: should be fixed as well, but I may have forgotten to execute something. The one marked with :three: is minor because this code block cannot be executed anyway.
I also found a few minor typos and typesetting comments.

#### p. 2
- "a code" -> "code"
- "to silently returns" -> "to silently return"
- bottom left "Using `a.num =` ..." does not look great. Maybe you can remove the line break before "Using".
- :three: first code block on the right (`function mul!!(a::Rational{S}, b::Rational{T})`) is missing `where {S,T}`

#### p. 3
- "always return" -> "always returns"
- list in front of "will be available": first three lines should probably end with a comma

#### p. 4
- :one: last code block: 3x `typeof` -> `eltype`
- :two: last code block: `@allocated buffered_operate!(buf, add_mul, c, A, b)` crashes
```julia
ERROR: MethodError: no method matching buffered_operate!(::BigInt, ::typeof(add_mul),
::Vector{BigInt}, ::Matrix{BigInt}, ::Vector{BigInt})

Closest candidates are:
 buffered_operate!(::Nothing, ::Any...)
  @ Main REPL[58]:1
 buffered_operate!(::BigInt, ::typeof(add_mul), ::BigInt, ::BigInt, ::BigInt)
  @ Main REPL[60]:1
```

#### p. 5
- :two: last code block: last two commands crash (probably same problem as above)
- reference [9]: capitalization of Julia and JuMP
