### 2024-04-08 ad06062
Add shortcut
