+++ paper/ref.bib
@@ -120,7 +120,7 @@
 
 @Conference{weisser2019polynomial,
   author    = {Weisser, Tillmann and Legat, Beno{\^\i}t and Coey, Chris and Kapelevich, Lea and Vielma, Juan Pablo},
-  title     = {Polynomial and Moment Optimization in Julia and JuMP},
+  title     = {{Polynomial and Moment Optimization in Julia and JuMP}},
   booktitle = {JuliaCon},
   year      = {2019},
   url       = {https://pretalx.com/juliacon2019/talk/QZBKAU/},
