### 2024-04-08 #93 @blegat
@schillic Thanks again for the review! the lines defining `buffered_operate!` are excerpt from the code of MA and do not have to be executed. I clarified it and added the namespace in front so that executing them does not cause issue. Let me know if the new version addresses your concerns
