### 2024-04-09 #93 @lucaferranti
@blegat to finalize the submission:

- could you check if some of the references with missing DOIs would have some DOI?

- Could you make a patch release with up-to-date version of software and paper and archive that to zenodo (if not already)

- post here the zenodo DOI
