### 2024-04-16 0a244ac
Move method and add tests
