+++ src/implementations/LinearAlgebra.jl
@@ -146,15 +146,6 @@ function operate!(op::AddSubMul, A::Array, x, y, args::Vararg{Any,N}) where {N}
     return operate!(op, A, x, *(y, args...))
 end
 
-function operate_to!(
-    output::AbstractArray,
-    op::Union{typeof(+),typeof(-)},
-    A::AbstractArray,
-    B::AbstractArray,
-)
-    return Base.broadcast!(op, output, A, B)
-end
-
 # Product
 
 function similar_array_type(
