### 2024-04-16 e34353d
Mutable sum of array to array
