+++ test/matmul.jl
@@ -119,6 +119,13 @@ end
             "Cannot sum or substract a matrix of axes `$(axes(B))` into matrix of axes `$(axes(A))`, expected axes `$(axes(B))`.",
         )
         @test_throws err MA.operate!(+, A, B)
+        output = zeros(2)
+        A = zeros(2, 1)
+        B = zeros(2, 1)
+        err = DimensionMismatch(
+            "Cannot sum or substract matrices of axes `$(axes(A))` and `$(axes(B))` into a matrix of axes `$(axes(output))`, expected axes `$(axes(B))`.",
+        )
+        @test_throws err MA.operate_to!(output, +, A, B)
     end
     @testset "unsupported_product" begin
         unsupported_product()
@@ -434,4 +441,18 @@ end
     @test A == reshape([1, 2], (2, 1))
     @test B == [1 2]
     @test D == B * A
+end
+
+function test_array_sum(::Type{T}) where {T}
+    x = zeros(T, 2)
+    y = copy(x)
+    z = copy(y)
+    alloc_test(() -> MA.operate!(+, y, z), 0)
+    alloc_test(() -> MA.add!!(y, z), 0)
+    alloc_test(() -> MA.operate_to!(x, +, y, z), 0)
+    alloc_test(() -> MA.add_to!!(x, y, z), 0)
+end
+
+@testset "Array sum" begin
+    test_array_sum(Int)
 end
