+++ test/matmul.jl
@@ -451,6 +451,7 @@ function test_array_sum(::Type{T}) where {T}
     alloc_test(() -> MA.add!!(y, z), 0)
     alloc_test(() -> MA.operate_to!(x, +, y, z), 0)
     alloc_test(() -> MA.add_to!!(x, y, z), 0)
+    return
 end
 
 @testset "Array sum" begin
