### 2024-04-18 #93 @blegat
> could you check if some of the references with missing DOIs would have some DOI?

I checked, they don't

> Could you make a patch release with up-to-date version of software and paper and archive that to zenodo (if not already)

Done

> post here the zenodo DOI

doi.org/10.5281/zenodo.10985724
