### 2024-04-18 #93 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
