### 2024-04-23 #93 @blegat
@lucaferranti @xuanxu @arfon I think there is an issue with the editorial bot.
The DOI used in his message https://github.com/JuliaCon/proceedings-review/issues/93#issuecomment-2064081312 is correct:
```
https://doi.org/10.21105/jcon.00093
```
but the DOI used in the badges of its last messages https://github.com/JuliaCon/proceedings-review/issues/93#issuecomment-2064107427 is not:
```
https://doi.org/10.21105/jcon.0093
```
There is also a `0` missing for the status badge url:
