### 2024-04-23 #93 @lucaferranti
Ok, so I took a look and, unless I am missing something, the only place with the incorrect DOI is the last comment above. Everywhere else, on the website, on the xml metadata, it seems to be correct.

So this should be only a minor inconvenience and should not mess up any metadata. For now, this is the correct link

Thanks for reporting this though, will look into it! The issue seems to be that one trailing zero went missing (0093 vs 00093)
