### 2024-04-23 #93 @blegat
Thanks for verifying that everything is correct. Indeed seems that only this generated message is affected
